REFACTOR: Remove all old references from Asset Issuer Actor Network Service, plus modified PluginRoot's name.
Associated Issue: #5720
