+++ DAP/plugin/actor_network_service/fermat-dap-plugin-actor-network-service-asset-issuer-bitdubai/src/main/java/com/bitdubai/fermat_dap_plugin/layer/actor/network/service/asset/issuer/developer/bitdubai/DeveloperBitDubai.java
@@ -15,7 +15,6 @@ import com.bitdubai.fermat_api.layer.all_definition.enums.Developers;
 import com.bitdubai.fermat_api.layer.all_definition.enums.TimeFrequency;
 import com.bitdubai.fermat_api.layer.all_definition.license.PluginLicensor;
 import com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.issuer.developer.bitdubai.version_1.AssetIssuerActorNetworkServicePluginRoot;
-import com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.issuer.developer.bitdubai.version_1.AssetIssuerActorNetworkServicePluginRootNew;
 
 
 /**
@@ -37,7 +36,7 @@ public class DeveloperBitDubai extends AbstractPluginDeveloper implements Plugin
         try {
 
 //            this.registerVersion(new AssetIssuerActorNetworkServicePluginRoot());
-            this.registerVersion(new AssetIssuerActorNetworkServicePluginRootNew());
+            this.registerVersion(new AssetIssuerActorNetworkServicePluginRoot());
 
         } catch (CantRegisterVersionException e) {
