+++ DAP/plugin/actor_network_service/fermat-dap-plugin-actor-network-service-asset-issuer-bitdubai/src/main/java/com/bitdubai/fermat_dap_plugin/layer/actor/network/service/asset/issuer/developer/bitdubai/version_1/AssetIssuerActorNetworkServicePluginRoot.java
@@ -8,16 +8,11 @@ package com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.issuer.
 
 import android.util.Base64;
 
-import com.bitdubai.fermat_api.CantStartAgentException;
 import com.bitdubai.fermat_api.CantStartPluginException;
-import com.bitdubai.fermat_api.FermatException;
-import com.bitdubai.fermat_api.layer.all_definition.common.system.annotations.NeededAddonReference;
-import com.bitdubai.fermat_api.layer.all_definition.common.system.annotations.NeededPluginReference;
 import com.bitdubai.fermat_api.layer.all_definition.common.system.utils.PluginVersionReference;
 import com.bitdubai.fermat_api.layer.all_definition.components.enums.PlatformComponentType;
 import com.bitdubai.fermat_api.layer.all_definition.components.interfaces.DiscoveryQueryParameters;
 import com.bitdubai.fermat_api.layer.all_definition.components.interfaces.PlatformComponentProfile;
-import com.bitdubai.fermat_api.layer.all_definition.crypto.asymmetric.ECCKeyPair;
 import com.bitdubai.fermat_api.layer.all_definition.developer.DatabaseManagerForDevelopers;
 import com.bitdubai.fermat_api.layer.all_definition.developer.DeveloperDatabase;
 import com.bitdubai.fermat_api.layer.all_definition.developer.DeveloperDatabaseTable;
@@ -25,110 +20,77 @@ import com.bitdubai.fermat_api.layer.all_definition.developer.DeveloperDatabaseT
 import com.bitdubai.fermat_api.layer.all_definition.developer.DeveloperObjectFactory;
 import com.bitdubai.fermat_api.layer.all_definition.developer.LogManagerForDevelopers;
 import com.bitdubai.fermat_api.layer.all_definition.enums.Actors;
-import com.bitdubai.fermat_api.layer.all_definition.enums.Addons;
 import com.bitdubai.fermat_api.layer.all_definition.enums.BlockchainNetworkType;
-import com.bitdubai.fermat_api.layer.all_definition.enums.Layers;
-import com.bitdubai.fermat_api.layer.all_definition.enums.Platforms;
 import com.bitdubai.fermat_api.layer.all_definition.enums.Plugins;
-import com.bitdubai.fermat_api.layer.all_definition.enums.ServiceStatus;
+import com.bitdubai.fermat_api.layer.all_definition.enums.SubAppsPublicKeys;
 import com.bitdubai.fermat_api.layer.all_definition.events.EventSource;
 import com.bitdubai.fermat_api.layer.all_definition.events.interfaces.FermatEvent;
-import com.bitdubai.fermat_api.layer.all_definition.events.interfaces.FermatEventListener;
+import com.bitdubai.fermat_api.layer.all_definition.exceptions.InvalidParameterException;
 import com.bitdubai.fermat_api.layer.all_definition.network_service.enums.NetworkServiceType;
-import com.bitdubai.fermat_api.layer.all_definition.network_service.interfaces.NetworkServiceConnectionManager;
 import com.bitdubai.fermat_api.layer.all_definition.util.Version;
 import com.bitdubai.fermat_api.layer.all_definition.util.XMLParser;
+import com.bitdubai.fermat_api.layer.osa_android.broadcaster.BroadcasterType;
 import com.bitdubai.fermat_api.layer.osa_android.database_system.Database;
-import com.bitdubai.fermat_api.layer.osa_android.database_system.PluginDatabaseSystem;
 import com.bitdubai.fermat_api.layer.osa_android.database_system.exceptions.CantCreateDatabaseException;
 import com.bitdubai.fermat_api.layer.osa_android.database_system.exceptions.CantOpenDatabaseException;
 import com.bitdubai.fermat_api.layer.osa_android.database_system.exceptions.DatabaseNotFoundException;
-import com.bitdubai.fermat_api.layer.osa_android.file_system.FileLifeSpan;
-import com.bitdubai.fermat_api.layer.osa_android.file_system.FilePrivacy;
-import com.bitdubai.fermat_api.layer.osa_android.file_system.PluginFileSystem;
-import com.bitdubai.fermat_api.layer.osa_android.file_system.PluginTextFile;
-import com.bitdubai.fermat_api.layer.osa_android.file_system.exceptions.CantCreateFileException;
-import com.bitdubai.fermat_api.layer.osa_android.file_system.exceptions.FileNotFoundException;
 import com.bitdubai.fermat_api.layer.osa_android.location_system.Location;
 import com.bitdubai.fermat_api.layer.osa_android.logger_system.LogLevel;
-import com.bitdubai.fermat_api.layer.osa_android.logger_system.LogManager;
 import com.bitdubai.fermat_bch_api.layer.crypto_vault.watch_only_vault.ExtendedPublicKey;
 import com.bitdubai.fermat_dap_api.layer.all_definition.enums.DAPMessageType;
 import com.bitdubai.fermat_dap_api.layer.all_definition.enums.EventType;
 import com.bitdubai.fermat_dap_api.layer.all_definition.events.ActorAssetIssuerCompleteRegistrationNotificationEvent;
-import com.bitdubai.fermat_dap_api.layer.all_definition.exceptions.CantHandleDapNewMessagesException;
+import com.bitdubai.fermat_dap_api.layer.all_definition.events.ActorAssetNetworkServicePendingNotificationEvent;
 import com.bitdubai.fermat_dap_api.layer.all_definition.network_service_message.DAPMessage;
 import com.bitdubai.fermat_dap_api.layer.all_definition.network_service_message.exceptions.CantGetDAPMessagesException;
-import com.bitdubai.fermat_dap_api.layer.all_definition.network_service_message.exceptions.CantSendMessageException;
-import com.bitdubai.fermat_dap_api.layer.all_definition.network_service_message.message.NetworkServiceMessage;
-import com.bitdubai.fermat_dap_api.layer.all_definition.network_service_message.message.NetworkServiceMessageAccept;
-import com.bitdubai.fermat_dap_api.layer.all_definition.network_service_message.message.NetworkServiceMessageDeny;
-import com.bitdubai.fermat_dap_api.layer.all_definition.network_service_message.message.NetworkServiceMessageRequest;
-import com.bitdubai.fermat_dap_api.layer.dap_actor.DAPActor;
+import com.bitdubai.fermat_dap_api.layer.all_definition.util.Validate;
 import com.bitdubai.fermat_dap_api.layer.dap_actor.asset_issuer.AssetIssuerActorRecord;
 import com.bitdubai.fermat_dap_api.layer.dap_actor.asset_issuer.interfaces.ActorAssetIssuer;
-import com.bitdubai.fermat_dap_api.layer.dap_actor.asset_user.interfaces.ActorAssetUser;
+import com.bitdubai.fermat_dap_api.layer.dap_actor_network_service.ActorAssetNetworkServiceRecord;
 import com.bitdubai.fermat_dap_api.layer.dap_actor_network_service.asset_issuer.exceptions.CantRegisterActorAssetIssuerException;
 import com.bitdubai.fermat_dap_api.layer.dap_actor_network_service.asset_issuer.exceptions.CantRequestListActorAssetIssuerRegisteredException;
 import com.bitdubai.fermat_dap_api.layer.dap_actor_network_service.asset_issuer.interfaces.AssetIssuerActorNetworkServiceManager;
+import com.bitdubai.fermat_dap_api.layer.dap_actor_network_service.enums.ActorAssetProtocolState;
+import com.bitdubai.fermat_dap_api.layer.dap_actor_network_service.enums.AssetNotificationDescriptor;
 import com.bitdubai.fermat_dap_api.layer.dap_actor_network_service.exceptions.CantAcceptConnectionActorAssetException;
 import com.bitdubai.fermat_dap_api.layer.dap_actor_network_service.exceptions.CantAskConnectionActorAssetException;
 import com.bitdubai.fermat_dap_api.layer.dap_actor_network_service.exceptions.CantCancelConnectionActorAssetException;
 import com.bitdubai.fermat_dap_api.layer.dap_actor_network_service.exceptions.CantConfirmActorAssetNotificationException;
+import com.bitdubai.fermat_dap_api.layer.dap_actor_network_service.exceptions.CantCreateActorAssetNotificationException;
 import com.bitdubai.fermat_dap_api.layer.dap_actor_network_service.exceptions.CantDenyConnectionActorAssetException;
 import com.bitdubai.fermat_dap_api.layer.dap_actor_network_service.exceptions.CantDisconnectConnectionActorAssetException;
 import com.bitdubai.fermat_dap_api.layer.dap_actor_network_service.exceptions.CantGetActorAssetNotificationException;
+import com.bitdubai.fermat_dap_api.layer.dap_actor_network_service.exceptions.CantUpdateRecordDataBaseException;
 import com.bitdubai.fermat_dap_api.layer.dap_actor_network_service.interfaces.ActorNotification;
-import com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.issuer.developer.bitdubai.version_1.agents.AssetIssuerActorNetworkServiceAgent;
-import com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.issuer.developer.bitdubai.version_1.communications.CommunicationNetworkServiceConnectionManager;
-import com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.issuer.developer.bitdubai.version_1.communications.CommunicationNetworkServiceLocal;
-import com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.issuer.developer.bitdubai.version_1.communications.CommunicationRegistrationProcessNetworkServiceAgent;
 import com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.issuer.developer.bitdubai.version_1.database.communications.AssetIssuerNetworkServiceDatabaseConstants;
 import com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.issuer.developer.bitdubai.version_1.database.communications.AssetIssuerNetworkServiceDatabaseFactory;
 import com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.issuer.developer.bitdubai.version_1.database.communications.AssetIssuerNetworkServiceDeveloperDatabaseFactory;
-import com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.issuer.developer.bitdubai.version_1.database.communications.OutgoingMessageDao;
-import com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.issuer.developer.bitdubai.version_1.event_handlers.ClientConnectionCloseNotificationEventHandler;
-import com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.issuer.developer.bitdubai.version_1.event_handlers.ClientConnectionLooseNotificationEventHandler;
-import com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.issuer.developer.bitdubai.version_1.event_handlers.ClientSuccessfullReconnectNotificationEventHandler;
-import com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.issuer.developer.bitdubai.version_1.event_handlers.CompleteComponentConnectionRequestNotificationEventHandler;
-import com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.issuer.developer.bitdubai.version_1.event_handlers.CompleteComponentRegistrationNotificationEventHandler;
-import com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.issuer.developer.bitdubai.version_1.event_handlers.CompleteRequestListComponentRegisteredNotificationEventHandler;
-import com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.issuer.developer.bitdubai.version_1.event_handlers.NewReceiveMessagesNotificationEventHandler;
-import com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.issuer.developer.bitdubai.version_1.event_handlers.NewSentMessagesNotificationEventHandler;
-import com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.issuer.developer.bitdubai.version_1.event_handlers.VPNConnectionCloseNotificationEventHandler;
+import com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.issuer.developer.bitdubai.version_1.database.communications.IncomingNotificationDao;
+import com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.issuer.developer.bitdubai.version_1.database.communications.OutgoingNotificationDao;
 import com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.issuer.developer.bitdubai.version_1.exceptions.CantInitializeTemplateNetworkServiceDatabaseException;
-import com.bitdubai.fermat_dap_api.layer.dap_actor_network_service.exceptions.CantReadRecordDataBaseException;
-import com.bitdubai.fermat_dap_api.layer.dap_actor_network_service.exceptions.CantUpdateRecordDataBaseException;
-import com.bitdubai.fermat_p2p_api.layer.all_definition.common.network_services.abstract_classes.AbstractNetworkService;
 import com.bitdubai.fermat_p2p_api.layer.all_definition.communication.commons.contents.FermatMessageCommunication;
-import com.bitdubai.fermat_p2p_api.layer.all_definition.communication.commons.contents.FermatMessageCommunicationFactory;
-import com.bitdubai.fermat_p2p_api.layer.all_definition.communication.enums.P2pEventType;
-import com.bitdubai.fermat_p2p_api.layer.all_definition.communication.events.ClientConnectionCloseNotificationEvent;
-import com.bitdubai.fermat_p2p_api.layer.all_definition.communication.events.VPNConnectionCloseNotificationEvent;
-import com.bitdubai.fermat_p2p_api.layer.p2p_communication.MessagesStatus;
-import com.bitdubai.fermat_p2p_api.layer.p2p_communication.WsCommunicationsCloudClientManager;
+import com.bitdubai.fermat_p2p_api.layer.all_definition.communication.network_services.base.AbstractNetworkServiceBase;
 import com.bitdubai.fermat_p2p_api.layer.p2p_communication.commons.client.CommunicationsClientConnection;
 import com.bitdubai.fermat_p2p_api.layer.p2p_communication.commons.contents.FermatMessage;
-import com.bitdubai.fermat_p2p_api.layer.p2p_communication.commons.enums.FermatMessageContentType;
 import com.bitdubai.fermat_p2p_api.layer.p2p_communication.commons.enums.FermatMessagesStatus;
 import com.bitdubai.fermat_p2p_api.layer.p2p_communication.commons.exceptions.CantRegisterComponentException;
 import com.bitdubai.fermat_p2p_api.layer.p2p_communication.commons.exceptions.CantRequestListException;
-import com.bitdubai.fermat_pip_api.layer.actor.exception.CantGetLogTool;
 import com.bitdubai.fermat_pip_api.layer.platform_service.error_manager.enums.UnexpectedPluginExceptionSeverity;
-import com.bitdubai.fermat_pip_api.layer.platform_service.error_manager.interfaces.ErrorManager;
-import com.bitdubai.fermat_pip_api.layer.platform_service.event_manager.interfaces.EventManager;
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
-import com.google.gson.JsonSyntaxException;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Timer;
+import java.util.TimerTask;
 import java.util.UUID;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.regex.Pattern;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+
+//import com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.issuer.developer.bitdubai.version_1.communications.CommunicationNetworkServiceLocal;
+//import com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.issuer.developer.bitdubai.version_1.database.communications.OutgoingMessageDao;
 
 /**
  * The Class <code>com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.issuer.developer.bitdubai.version_1.AssetIssuerActorNetworkServicePluginRoot</code> is
@@ -139,67 +101,37 @@ import java.util.regex.Pattern;
  *
  * @version 1.0
  */
-public class AssetIssuerActorNetworkServicePluginRoot extends AbstractNetworkService implements
+public class AssetIssuerActorNetworkServicePluginRoot extends AbstractNetworkServiceBase implements
         AssetIssuerActorNetworkServiceManager,
         DatabaseManagerForDevelopers,
         LogManagerForDevelopers {
-
-    @NeededAddonReference(platform = Platforms.OPERATIVE_SYSTEM_API, layer = Layers.SYSTEM, addon = Addons.PLUGIN_FILE_SYSTEM)
-    protected PluginFileSystem pluginFileSystem;
-
-    @NeededAddonReference(platform = Platforms.OPERATIVE_SYSTEM_API, layer = Layers.SYSTEM, addon = Addons.PLUGIN_DATABASE_SYSTEM)
-    private PluginDatabaseSystem pluginDatabaseSystem;
-
-    @NeededAddonReference(platform = Platforms.OPERATIVE_SYSTEM_API, layer = Layers.SYSTEM, addon = Addons.LOG_MANAGER)
-    private LogManager logManager;
-
-    @NeededAddonReference(platform = Platforms.PLUG_INS_PLATFORM, layer = Layers.PLATFORM_SERVICE, addon = Addons.ERROR_MANAGER)
-    private ErrorManager errorManager;
-
-    @NeededAddonReference(platform = Platforms.PLUG_INS_PLATFORM, layer = Layers.PLATFORM_SERVICE, addon = Addons.EVENT_MANAGER)
-    private EventManager eventManager;
-
-    @NeededPluginReference(platform = Platforms.COMMUNICATION_PLATFORM, layer = Layers.COMMUNICATION, plugin = Plugins.WS_CLOUD_CLIENT)
-    private WsCommunicationsCloudClientManager wsCommunicationsCloudClientManager;
 
     /**
      * Represent the dataBase
      */
     private Database dataBase;
 
-    /**
-     * Represent the EVENT_SOURCE
-     */
-    public final static EventSource EVENT_SOURCE = EventSource.NETWORK_SERVICE_ACTOR_ASSET_ISSUER;
-
     protected final static String DAP_IMG_ISSUER = "DAP_IMG_ISSUER";
 
-    List<FermatEventListener> listenersAdded = new ArrayList<>();
-
-    /**
-     * DealsWithLogger interface member variable
-     */
-    static Map<String, LogLevel> newLoggingLevel = new HashMap<>();
-
-    /**
-     * Represent the remoteNetworkServicesRegisteredList
-     */
-    private List<PlatformComponentProfile> remoteNetworkServicesRegisteredList;
+//    /**
+//     * DealsWithLogger interface member variable
+//     */
+//    static Map<String, LogLevel> newLoggingLevel = new HashMap<>();
 
     /**
      * Represent the communicationNetworkServiceDeveloperDatabaseFactory
      */
-    private AssetIssuerNetworkServiceDeveloperDatabaseFactory communicationNetworkServiceDeveloperDatabaseFactory;
+    private AssetIssuerNetworkServiceDeveloperDatabaseFactory assetIssuerNetworkServiceDeveloperDatabaseFactory;
 
     /**
      * Represent the communicationRegistrationProcessNetworkServiceAgent
      */
-    private CommunicationRegistrationProcessNetworkServiceAgent communicationRegistrationProcessNetworkServiceAgent;
+//    private CommunicationRegistrationProcessNetworkServiceAgent communicationRegistrationProcessNetworkServiceAgent;
 
     /**
      * Represent the communicationNetworkServiceConnectionManager
      */
-    private CommunicationNetworkServiceConnectionManager communicationNetworkServiceConnectionManager;
+//    private CommunicationNetworkServiceConnectionManager communicationNetworkServiceConnectionManager;
 
     /**
      * Represent the actorAssetIssuerRegisteredList
@@ -211,15 +143,21 @@ public class AssetIssuerActorNetworkServicePluginRoot extends AbstractNetworkSer
      */
     private List<PlatformComponentProfile> actorAssetIssuerPendingToRegistration;
 
-    private int createactorAgentnum = 0;
-
     /**
-     * Represent the flag to start only once
+     * DAO
      */
-    private AtomicBoolean flag = new AtomicBoolean(false);
+    private IncomingNotificationDao incomingNotificationsDao;
+    private OutgoingNotificationDao outgoingNotificationDao;
 
-    private AssetIssuerActorNetworkServiceAgent assetIssuerActorNetworkServiceAgent;
+    private long reprocessTimer = 300000; //five minutes
 
+    private Timer timer = new Timer();
+    /**
+     * Executor
+     */
+    ExecutorService executorService;
+
+//    private AssetIssuerActorNetworkServiceAgent assetIssuerActorNetworkServiceAgent;
 
     /**
      * Constructor
@@ -227,201 +165,392 @@ public class AssetIssuerActorNetworkServicePluginRoot extends AbstractNetworkSer
     public AssetIssuerActorNetworkServicePluginRoot() {
         super(
                 new PluginVersionReference(new Version()),
+                EventSource.NETWORK_SERVICE_ACTOR_ASSET_ISSUER,
                 PlatformComponentType.NETWORK_SERVICE,
                 NetworkServiceType.ASSET_ISSUER_ACTOR,
                 "Actor Network Service Asset Issuer",
-                "ActorNetworkServiceAssetIssuer",
-                null,
-                EventSource.NETWORK_SERVICE_ACTOR_ASSET_ISSUER
+                null
         );
-        this.listenersAdded = new ArrayList<>();
-
         this.actorAssetIssuerRegisteredList = new ArrayList<>();
         this.actorAssetIssuerPendingToRegistration = new ArrayList<>();
     }
 
     @Override
-    public List<String> getClassesFullPath() {
-        List<String> returnedClasses = new ArrayList<String>();
-        returnedClasses.add("com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.issuer.developer.bitdubai.version_1.AssetIssuerActorNetworkServicePluginRoot");
-        /**
-         * I return the values.
-         */
-        return returnedClasses;
-    }
-
-    @Override
-    public void setLoggingLevelPerClass(Map<String, LogLevel> newLoggingLevel) {
-        /**
-         * Modify by Manuel on 25/07/2015
-         * I will wrap all this method within a try, I need to catch any generic java Exception
-         */
+    protected void onStart() {
         try {
+        /*
+         * Initialize the data base
+         */
+            initializeDb();
+        /*
+         * Initialize Developer Database Factory
+         */
+            assetIssuerNetworkServiceDeveloperDatabaseFactory = new AssetIssuerNetworkServiceDeveloperDatabaseFactory(pluginDatabaseSystem, pluginId);
+            assetIssuerNetworkServiceDeveloperDatabaseFactory.initializeDatabase();
 
-            /**
-             * I will check the current values and update the LogLevel in those which is different
-             */
-            for (Map.Entry<String, LogLevel> pluginPair : newLoggingLevel.entrySet()) {
-                /**
-                 * if this path already exists in the Root.bewLoggingLevel I'll update the value, else, I will put as new
-                 */
-                if (AssetIssuerActorNetworkServicePluginRoot.newLoggingLevel.containsKey(pluginPair.getKey())) {
-                    AssetIssuerActorNetworkServicePluginRoot.newLoggingLevel.remove(pluginPair.getKey());
-                    AssetIssuerActorNetworkServicePluginRoot.newLoggingLevel.put(pluginPair.getKey(), pluginPair.getValue());
-                } else {
-                    AssetIssuerActorNetworkServicePluginRoot.newLoggingLevel.put(pluginPair.getKey(), pluginPair.getValue());
-                }
-            }
+            //DAO
+            incomingNotificationsDao = new IncomingNotificationDao(dataBase, this.pluginFileSystem, this.pluginId);
 
-        } catch (Exception exception) {
-            FermatException e = new CantGetLogTool(CantGetLogTool.DEFAULT_MESSAGE, FermatException.wrapException(exception), "setLoggingLevelPerClass: " + AssetIssuerActorNetworkServicePluginRoot.newLoggingLevel, "Check the cause");
-            // this.errorManager.reportUnexpectedAddonsException(Addons.EXTRA_USER, UnexpectedAddonsExceptionSeverity.DISABLES_THIS_ADDONS, e);
+            outgoingNotificationDao = new OutgoingNotificationDao(dataBase, this.pluginFileSystem, this.pluginId);
+
+            executorService = Executors.newFixedThreadPool(2);
+
+            // change message state to process again first time
+            reprocessPendingMessage();
+
+            //declare a schedule to process waiting request message
+            this.startTimer();
+
+        } catch (Exception e) {
+            e.printStackTrace();
         }
     }
 
     @Override
-    public synchronized void start() throws CantStartPluginException {
+    public void stop() {
+        super.stop();
+        executorService.shutdownNow();
+    }
 
-        if (!flag.getAndSet(true)) {
-            if (this.serviceStatus != ServiceStatus.STARTING) {
-                serviceStatus = ServiceStatus.STARTING;
+    @Override
+    public void onNewMessagesReceive(FermatMessage newFermatMessageReceive) {
+        try {
+            System.out.println("ACTOR ASSET ISSUER MENSAJE ENTRANTE A JSON: " + newFermatMessageReceive.toJson());
 
-                System.out.println("Star Plugin AssetIssuerActorNetworkService");
-                logManager.log(AssetIssuerActorNetworkServicePluginRoot.getLogLevelByClass(this.getClass().getName()), "AssetIssuerActorNetworkService - Starting", "AssetIssuerActorNetworkServicePluginRoot - Starting", "AssetIssuerActorNetworkServicePluginRoot - Starting");
+            ActorAssetNetworkServiceRecord assetUserNetworkServiceRecord = ActorAssetNetworkServiceRecord.fronJson(newFermatMessageReceive.getContent());
 
-        /*
-         * Validate required resources
-         */
-                validateInjectedResources();
+            switch (assetUserNetworkServiceRecord.getAssetNotificationDescriptor()) {
+                case ASKFORCONNECTION:
+                    System.out.println("ACTOR ASSET ISSUER MENSAJE LLEGO: " + assetUserNetworkServiceRecord.getActorSenderAlias());
+                    assetUserNetworkServiceRecord.changeState(ActorAssetProtocolState.PROCESSING_RECEIVE);
+                    System.out.println("ACTOR ASSET ISSUER REGISTRANDO EN INCOMING NOTIFICATION DAO");
+                    assetUserNetworkServiceRecord.setFlagRead(false);
+                    incomingNotificationsDao.createNotification(assetUserNetworkServiceRecord);
 
-                try {
-        /*
-         *  Create a new key pair for this execution
-         */
-                    initializeClientIdentity();
-            /*
-            * Initialize the data base
-            */
-                    initializeDb();
+                    //NOTIFICATION LAUNCH
+                    launchNotificationActorAsset();
+                    broadcaster.publish(BroadcasterType.NOTIFICATION_SERVICE, SubAppsPublicKeys.DAP_COMMUNITY_ISSUER.getCode(), "EXTENDED-REQUEST_" + assetUserNetworkServiceRecord.getActorSenderPublicKey());
 
-                        /*
-             * Initialize Developer Database Factory
-             */
-                    communicationNetworkServiceDeveloperDatabaseFactory = new AssetIssuerNetworkServiceDeveloperDatabaseFactory(pluginDatabaseSystem, pluginId);
-                    communicationNetworkServiceDeveloperDatabaseFactory.initializeDatabase();
-                        /*
-             * Initialize listeners
-             */
-                    initializeListener();
+                    respondReceiveAndDoneCommunication(assetUserNetworkServiceRecord);
+                    break;
 
-            /*
-             * Initialize connection manager
-             */
-                    initializeCommunicationNetworkServiceConnectionManager();
+                case ACCEPTED:
+                    //TODO: ver si me conviene guardarlo en el outogoing DAO o usar el incoming para las que llegan directamente
+                    assetUserNetworkServiceRecord.changeDescriptor(AssetNotificationDescriptor.ACCEPTED);
+                    assetUserNetworkServiceRecord.changeState(ActorAssetProtocolState.DONE);
+                    outgoingNotificationDao.update(assetUserNetworkServiceRecord);
 
-            /*
-             * Verify if the communication cloud client is active
-             */
-                    if (!wsCommunicationsCloudClientManager.isDisable()) {
+                    //create incoming notification
+                    assetUserNetworkServiceRecord.changeState(ActorAssetProtocolState.PROCESSING_RECEIVE);
+                    assetUserNetworkServiceRecord.setFlagRead(false);
+                    incomingNotificationsDao.createNotification(assetUserNetworkServiceRecord);
+                    System.out.println("ACTOR ASSET ISSUER MENSAJE ACCEPTED LLEGÓ: " + assetUserNetworkServiceRecord.getActorSenderAlias());
 
-                /*
-                 * Construct my profile and register me
-                 */
-                        PlatformComponentProfile platformComponentProfilePluginRoot = wsCommunicationsCloudClientManager.getCommunicationsCloudClientConnection().constructPlatformComponentProfileFactory(getIdentityPublicKey(),
-                                getAlias().toLowerCase(),
-                                getName(),
-                                getNetworkServiceType(),
-                                getPlatformComponentType(),
-                                getExtraData());
+                    //NOTIFICATION LAUNCH
+                    launchNotificationActorAsset();
+                    respondReceiveAndDoneCommunication(assetUserNetworkServiceRecord);
+                    break;
 
-                        setPlatformComponentProfilePluginRoot(platformComponentProfilePluginRoot);
+                case RECEIVED:
+                    //launchIncomingRequestConnectionNotificationEvent(actorNetworkServiceRecord);
+                    System.out.println("ACTOR ASSET ISSUER THE RECORD WAS CHANGE TO THE STATE OF DELIVERY: " + assetUserNetworkServiceRecord.getActorSenderAlias());
+                    if (assetUserNetworkServiceRecord.getResponseToNotificationId() != null)
+                        outgoingNotificationDao.changeProtocolState(assetUserNetworkServiceRecord.getResponseToNotificationId(), ActorAssetProtocolState.DONE);
 
-                /*
-                 * Initialize the agent and start
-                 */
-                        communicationRegistrationProcessNetworkServiceAgent = new CommunicationRegistrationProcessNetworkServiceAgent(this, wsCommunicationsCloudClientManager);
-                        communicationRegistrationProcessNetworkServiceAgent.start();
+                    // close connection, sender is the destination
+                    System.out.println("ACTOR ASSET ISSUER THE CONNECTION WAS CHANGE TO DONE" + assetUserNetworkServiceRecord.getActorSenderAlias());
+
+                    getCommunicationNetworkServiceConnectionManager().closeConnection(assetUserNetworkServiceRecord.getActorSenderPublicKey());
+                    System.out.println("ACTOR ASSET ISSUER THE CONNECTION WAS CLOSED AND THE AWAITING POOL CLEARED." + assetUserNetworkServiceRecord.getActorSenderAlias());
+                    break;
+
+                case DENIED:
+                    assetUserNetworkServiceRecord.changeDescriptor(AssetNotificationDescriptor.DENIED);
+                    assetUserNetworkServiceRecord.changeState(ActorAssetProtocolState.DONE);
+                    outgoingNotificationDao.update(assetUserNetworkServiceRecord);
+
+                    assetUserNetworkServiceRecord.changeState(ActorAssetProtocolState.PROCESSING_RECEIVE);
+                    assetUserNetworkServiceRecord.setFlagRead(false);
+                    incomingNotificationsDao.createNotification(assetUserNetworkServiceRecord);
+                    System.out.println("ACTOR ASSET ISSUER MENSAJE DENIED LLEGÓ: " + assetUserNetworkServiceRecord.getActorDestinationPublicKey());
+
+                    //NOTIFICATION LAUNCH
+                    launchNotificationActorAsset();
+                    respondReceiveAndDoneCommunication(assetUserNetworkServiceRecord);
+                    break;
+
+                case DISCONNECTED:
+                    assetUserNetworkServiceRecord.changeDescriptor(AssetNotificationDescriptor.DISCONNECTED);
+                    assetUserNetworkServiceRecord.changeState(ActorAssetProtocolState.DONE);
+                    outgoingNotificationDao.update(assetUserNetworkServiceRecord);
+
+                    assetUserNetworkServiceRecord.changeState(ActorAssetProtocolState.PROCESSING_RECEIVE);
+                    assetUserNetworkServiceRecord.setFlagRead(false);
+                    incomingNotificationsDao.createNotification(assetUserNetworkServiceRecord);
+                    System.out.println("ACTOR ASSET ISSUER MENSAJE DISCONNECTED LLEGÓ: " + assetUserNetworkServiceRecord.getActorSenderAlias());
+
+                    //NOTIFICATION LAUNCH
+                    launchNotificationActorAsset();
+                    respondReceiveAndDoneCommunication(assetUserNetworkServiceRecord);
+                    break;
+                case EXTENDED_KEY:
+                    if (assetUserNetworkServiceRecord.getMessageXML() == null) {
+                        System.out.println("ACTOR ASSET ISSUER REQUEST EXTENDED-KEY: " + assetUserNetworkServiceRecord.getActorSenderAlias());
+                        System.out.println("ACTOR ASSET ISSUER REGISTRANDO EN INCOMING NOTIFICATION DAO: REQUEST EXTENDED-KEY");
                     }
+                    assetUserNetworkServiceRecord.changeState(ActorAssetProtocolState.PROCESSING_RECEIVE);
+                    assetUserNetworkServiceRecord.setFlagRead(false);
+                    incomingNotificationsDao.createNotification(assetUserNetworkServiceRecord);
+                    //NOTIFICATION LAUNCH
+                    launchNotificationActorAsset();
+//                    broadcaster.publish(BroadcasterType.NOTIFICATION_SERVICE, DAPPublicKeys.DAP_COMMUNITY_ISSUER.getCode(), "CONNECTION-REQUEST_" + assetUserNetworkServiceRecord.getActorSenderPublicKey());
 
-                    this.serviceStatus = ServiceStatus.STARTED;
-                } catch (CantInitializeTemplateNetworkServiceDatabaseException exception) {
+                    respondReceiveAndDoneCommunication(assetUserNetworkServiceRecord);
+                    break;
 
-                    StringBuffer contextBuffer = new StringBuffer();
-                    contextBuffer.append("Plugin ID: " + pluginId);
-                    contextBuffer.append(CantStartPluginException.CONTEXT_CONTENT_SEPARATOR);
-                    contextBuffer.append("Database Name: " + AssetIssuerNetworkServiceDatabaseConstants.DATA_BASE_NAME);
-
-                    String context = contextBuffer.toString();
-                    String possibleCause = "The Asset Issuer Actor Network Service Database triggered an unexpected problem that wasn't able to solve by itself";
-                    CantStartPluginException pluginStartException = new CantStartPluginException(CantStartPluginException.DEFAULT_MESSAGE, exception, context, possibleCause);
-
-                    errorManager.reportUnexpectedPluginException(Plugins.BITDUBAI_DAP_ASSET_ISSUER_ACTOR_NETWORK_SERVICE, UnexpectedPluginExceptionSeverity.DISABLES_THIS_PLUGIN, pluginStartException);
-
-                    throw pluginStartException;
-                }
+                default:
+                    break;
             }
+
+        } catch (Exception e) {
+            //quiere decir que no estoy reciviendo metadata si no una respuesta
+            e.printStackTrace();
+
+        }
+
+        System.out.println("Actor Asset Issuer Llegaron mensajes!!!!");
+
+        try {
+            getCommunicationNetworkServiceConnectionManager().getIncomingMessageDao().markAsRead(newFermatMessageReceive);
+        } catch (com.bitdubai.fermat_p2p_api.layer.all_definition.communication.network_services.exceptions.CantUpdateRecordDataBaseException e) {
+            e.printStackTrace();
         }
     }
 
-    private void initializeClientIdentity() throws CantStartPluginException {
-
-        System.out.println("Calling the method - initializeClientIdentity() ");
-
+    @Override
+    public void onSentMessage(FermatMessage messageSent) {
         try {
+            ActorAssetNetworkServiceRecord assetUserNetworkServiceRecord = ActorAssetNetworkServiceRecord.fronJson(messageSent.getContent());
 
-            System.out.println("Loading clientIdentity");
-
-             /*
-              * Load the file with the clientIdentity
-              */
-            PluginTextFile pluginTextFile = pluginFileSystem.getTextFile(pluginId, "private", "clientIdentity", FilePrivacy.PRIVATE, FileLifeSpan.PERMANENT);
-            String content = pluginTextFile.getContent();
-
-            //System.out.println("content = "+content);
-
-            identity = new ECCKeyPair(content);
-
-        } catch (FileNotFoundException e) {
-
-            /*
-             * The file no exist may be the first time the plugin is running on this device,
-             * We need to create the new clientIdentity
-             */
-            try {
-
-                System.out.println("No previous clientIdentity finder - Proceed to create new one");
-
-                /*
-                 * Create the new clientIdentity
-                 */
-                identity = new ECCKeyPair();
-
-                /*
-                 * save into the file
-                 */
-                PluginTextFile pluginTextFile = pluginFileSystem.createTextFile(pluginId, "private", "clientIdentity", FilePrivacy.PRIVATE, FileLifeSpan.PERMANENT);
-                pluginTextFile.setContent(identity.getPrivateKey());
-                pluginTextFile.persistToMedia();
-
-            } catch (Exception exception) {
-                /*
-                 * The file cannot be created. I can not handle this situation.
-                 */
-                errorManager.reportUnexpectedPluginException(Plugins.BITDUBAI_DAP_ASSET_ISSUER_ACTOR_NETWORK_SERVICE, UnexpectedPluginExceptionSeverity.DISABLES_SOME_FUNCTIONALITY_WITHIN_THIS_PLUGIN, exception);
-                throw new CantStartPluginException(exception.getLocalizedMessage());
+            if (assetUserNetworkServiceRecord.getActorAssetProtocolState() == ActorAssetProtocolState.DONE) {
+                // close connection, sender is the destination
+                System.out.println("ACTOR ASSET ISSUER CERRANDO LA CONEXION DEL HANDLE NEW SENT MESSAGE NOTIFICATION");
+                //   communicationNetworkServiceConnectionManager.closeConnection(actorNetworkServiceRecord.getActorDestinationPublicKey());
+//                assetUserActorNetworkServiceAgent.getPoolConnectionsWaitingForResponse().remove(assetUserNetworkServiceRecord.getActorDestinationPublicKey());
             }
 
+            //done message type receive
+            if (assetUserNetworkServiceRecord.getAssetNotificationDescriptor() == AssetNotificationDescriptor.RECEIVED) {
+                assetUserNetworkServiceRecord.setActorAssetProtocolState(ActorAssetProtocolState.DONE);
+                outgoingNotificationDao.update(assetUserNetworkServiceRecord);
+//                assetUserActorNetworkServiceAgent.getPoolConnectionsWaitingForResponse().remove(assetUserNetworkServiceRecord.getActorDestinationPublicKey());
+            }
+            System.out.println("SALIENDO DEL HANDLE ACTOR ASSET ISSUER NEW SENT MESSAGE NOTIFICATION");
 
-        } catch (CantCreateFileException cantCreateFileException) {
-
-            /*
-             * The file cannot be load. I can not handle this situation.
-             */
-            errorManager.reportUnexpectedPluginException(Plugins.BITDUBAI_DAP_ASSET_ISSUER_ACTOR_NETWORK_SERVICE, UnexpectedPluginExceptionSeverity.DISABLES_SOME_FUNCTIONALITY_WITHIN_THIS_PLUGIN, cantCreateFileException);
-            throw new CantStartPluginException(cantCreateFileException.getLocalizedMessage());
+        } catch (Exception e) {
+            //quiere decir que no estoy reciviendo metadata si no una respuesta
+            System.out.println("EXCEPCION DENTRO DEL PROCCESS EVENT IN ACTOR ASSET ISSUER");
+            e.printStackTrace();
 
         }
+    }
 
+    @Override
+    protected void onNetworkServiceRegistered() {
+        try {
+            //TODO Test this functionality
+            for (PlatformComponentProfile platformComponentProfile : actorAssetIssuerPendingToRegistration) {
+                wsCommunicationsCloudClientManager.getCommunicationsCloudClientConnection().registerComponentForCommunication(getNetworkServiceProfile().getNetworkServiceType(), platformComponentProfile);
+                System.out.println("AssetIssuerActorNetworkServicePluginRoot - Trying to register to: " + platformComponentProfile.getAlias());
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+
+    @Override
+    protected void onFailureComponentConnectionRequest(PlatformComponentProfile remoteParticipant) {
+        //I check my time trying to send the message
+        checkFailedDeliveryTime(remoteParticipant.getIdentityPublicKey());
+    }
+
+//    @Override
+//    public PlatformComponentProfile getProfileSenderToRequestConnection(String identityPublicKeySender) {
+//        try {
+//
+//            Actors actors = outgoingNotificationDao.getActorTypeFromRequest(identityPublicKeySender);
+//
+//            return wsCommunicationsCloudClientManager.getCommunicationsCloudClientConnection()
+//                    .constructPlatformComponentProfileFactory(identityPublicKeySender,
+//                            "sender_alias",
+//                            "sender_name",
+//                            NetworkServiceType.UNDEFINED,
+//                            platformComponentTypeSelectorByActorType(actors),
+//                            "");
+//        } catch (Exception e) {
+//            e.printStackTrace();
+//        }
+//        return null;
+//    }
+//
+//    @Override
+//    public PlatformComponentProfile getProfileDestinationToRequestConnection(String identityPublicKeyDestination) {
+//        try {
+//
+//            Actors actors = outgoingNotificationDao.getActorTypeToRequest(identityPublicKeyDestination);
+//
+//            return wsCommunicationsCloudClientManager.getCommunicationsCloudClientConnection()
+//                    .constructPlatformComponentProfileFactory(identityPublicKeyDestination,
+//                            "destination_alias",
+//                            "destination_name",
+//                            NetworkServiceType.UNDEFINED,
+//                            platformComponentTypeSelectorByActorType(actors),
+//                            "");
+//        } catch (Exception e) {
+//            e.printStackTrace();
+//        }
+//        return null;
+//    }
+
+    private void reprocessPendingMessage() {
+        try {
+            outgoingNotificationDao.changeStatusNotSentMessage();
+
+            List<ActorAssetNetworkServiceRecord> lstActorRecord = outgoingNotificationDao.
+                    listRequestsByProtocolStateAndNotDone(
+                            ActorAssetProtocolState.PROCESSING_SEND
+                    );
+
+            for (final ActorAssetNetworkServiceRecord cpr : lstActorRecord) {
+
+                executorService.submit(new Runnable() {
+                    @Override
+                    public void run() {
+                        try {
+                            sendNewMessage(
+//                                    getProfileSenderToRequestConnection(cpr.getActorSenderPublicKey()),
+//                                    getProfileDestinationToRequestConnection(cpr.getActorDestinationPublicKey()),
+//                                    cpr.toJson());
+
+                                    getProfileSenderToRequestConnection(
+                                            cpr.getActorSenderPublicKey(),
+                                            NetworkServiceType.UNDEFINED,
+                                            platformComponentTypeSelectorByActorType(cpr.getActorSenderType())
+                                    ),
+                                    getProfileDestinationToRequestConnection(
+                                            cpr.getActorDestinationPublicKey(),
+                                            NetworkServiceType.UNDEFINED,
+                                            platformComponentTypeSelectorByActorType(cpr.getActorDestinationType())
+                                    ),
+                                    cpr.toJson());
+                        } catch (com.bitdubai.fermat_p2p_api.layer.all_definition.communication.network_services.exceptions.CantSendMessageException | InvalidParameterException e) {
+                            e.printStackTrace();
+                        }
+                    }
+                });
+            }
+        } catch (CantGetActorAssetNotificationException e) {
+            System.out.println("ACTOR ASSET USER NS EXCEPCION REPROCESANDO MESSAGEs");
+            e.printStackTrace();
+        } catch (Exception e) {
+            System.out.println("ACTOR ASSET USER NS EXCEPCION REPROCESANDO MESSAGEs");
+            e.printStackTrace();
+        }
+    }
+
+    @Override
+    protected void reprocessMessages() {
+//        try {
+//            outgoingNotificationDao.changeStatusNotSentMessage();
+//        } catch (CantGetActorAssetNotificationException e) {
+//            System.out.println("INTRA USER NS EXCEPCION REPROCESANDO MESSAGEs");
+//            e.printStackTrace();
+//        } catch (Exception e) {
+//            System.out.println("INTRA USER NS EXCEPCION REPROCESANDO MESSAGEs");
+//            e.printStackTrace();
+//        }
+    }
+
+    @Override
+    protected void reprocessMessages(String identityPublicKey) {
+//        try {
+//            outgoingNotificationDao.changeStatusNotSentMessage(identityPublicKey);
+//        } catch (CantGetActorAssetNotificationException e) {
+//            System.out.println("ACTOR ASSET USER NS EXCEPCION REPROCESANDO MESSAGEs");
+//            e.printStackTrace();
+//        } catch (Exception e) {
+//            System.out.println("ACTOR ASSET USER NS EXCEPCION REPROCESANDO MESSAGEs");
+//            e.printStackTrace();
+//        }
+    }
+
+    private PlatformComponentType platformComponentTypeSelectorByActorType(final Actors type) throws InvalidParameterException {
+
+        switch (type) {
+            case DAP_ASSET_ISSUER:
+                return PlatformComponentType.ACTOR_ASSET_ISSUER;
+            case DAP_ASSET_USER:
+                return PlatformComponentType.ACTOR_ASSET_USER;
+            case DAP_ASSET_REDEEM_POINT:
+                return PlatformComponentType.ACTOR_ASSET_REDEEM_POINT;
+
+            default:
+                throw new InvalidParameterException(
+                        " actor type: " + type.name() + "  type-code: " + type.getCode(),
+                        " type of actor not expected."
+                );
+        }
+    }
+
+    private void launchNotificationActorAsset() {
+        FermatEvent fermatEvent = eventManager.getNewEvent(EventType.ACTOR_ASSET_NETWORK_SERVICE_NEW_NOTIFICATIONS);
+        fermatEvent.setSource(EventSource.NETWORK_SERVICE_ACTOR_ASSET_ISSUER);
+        ActorAssetNetworkServicePendingNotificationEvent actorAssetRequestConnectionEvent = (ActorAssetNetworkServicePendingNotificationEvent) fermatEvent;
+        eventManager.raiseEvent(actorAssetRequestConnectionEvent);
+    }
+
+    private ActorAssetNetworkServiceRecord swapActor(ActorAssetNetworkServiceRecord assetUserNetworkServiceRecord) {
+        // swap actor
+        String actorDestination = assetUserNetworkServiceRecord.getActorDestinationPublicKey();
+        Actors actorsType = assetUserNetworkServiceRecord.getActorDestinationType();
+
+        assetUserNetworkServiceRecord.setActorDestinationPublicKey(assetUserNetworkServiceRecord.getActorSenderPublicKey());
+        assetUserNetworkServiceRecord.setActorDestinationType(assetUserNetworkServiceRecord.getActorSenderType());
+
+        assetUserNetworkServiceRecord.setActorSenderPublicKey(actorDestination);
+        assetUserNetworkServiceRecord.setActorSenderType(actorsType);
+
+        return assetUserNetworkServiceRecord;
+    }
+
+    // respond receive and done notification
+    private void respondReceiveAndDoneCommunication(ActorAssetNetworkServiceRecord assetUserNetworkServiceRecord) {
+
+        assetUserNetworkServiceRecord = swapActor(assetUserNetworkServiceRecord);
+        try {
+            UUID newNotificationID = UUID.randomUUID();
+            long currentTime = System.currentTimeMillis();
+            ActorAssetProtocolState actorAssetProtocolState = ActorAssetProtocolState.PROCESSING_SEND;
+            assetUserNetworkServiceRecord.changeDescriptor(AssetNotificationDescriptor.RECEIVED);
+            outgoingNotificationDao.createNotification(
+                    newNotificationID,
+                    assetUserNetworkServiceRecord.getActorSenderPublicKey(),
+                    assetUserNetworkServiceRecord.getActorSenderType(),
+                    assetUserNetworkServiceRecord.getActorDestinationPublicKey(),
+                    assetUserNetworkServiceRecord.getActorSenderAlias(),
+                    assetUserNetworkServiceRecord.getActorSenderProfileImage(),
+                    assetUserNetworkServiceRecord.getActorDestinationType(),
+                    assetUserNetworkServiceRecord.getAssetNotificationDescriptor(),
+                    currentTime,
+                    actorAssetProtocolState,
+                    false,
+                    1,
+                    assetUserNetworkServiceRecord.getBlockchainNetworkType(),
+                    assetUserNetworkServiceRecord.getId(), null
+            );
+        } catch (CantCreateActorAssetNotificationException e) {
+            e.printStackTrace();
+        }
     }
 
     /**
@@ -442,7 +571,7 @@ public class AssetIssuerActorNetworkServicePluginRoot extends AbstractNetworkSer
             /*
              * The database exists but cannot be open. I can not handle this situation.
              */
-            //errorManager.reportUnexpectedPluginException(Plugins.BITDUBAI_DAP_ASSET_USER_ACTOR_NETWORK_SERVICE, UnexpectedPluginExceptionSeverity.DISABLES_THIS_PLUGIN, cantOpenDatabaseException);
+            errorManager.reportUnexpectedPluginException(Plugins.BITDUBAI_DAP_ASSET_ISSUER_ACTOR_NETWORK_SERVICE, UnexpectedPluginExceptionSeverity.DISABLES_THIS_PLUGIN, cantOpenDatabaseException);
             throw new CantInitializeTemplateNetworkServiceDatabaseException(cantOpenDatabaseException.getLocalizedMessage());
 
         } catch (DatabaseNotFoundException e) {
@@ -451,166 +580,86 @@ public class AssetIssuerActorNetworkServicePluginRoot extends AbstractNetworkSer
              * The database no exist may be the first time the plugin is running on this device,
              * We need to create the new database
              */
-            AssetIssuerNetworkServiceDatabaseFactory communicationNetworkServiceDatabaseFactory = new AssetIssuerNetworkServiceDatabaseFactory(pluginDatabaseSystem);
+            AssetIssuerNetworkServiceDatabaseFactory assetIssuerNetworkServiceDatabaseFactory = new AssetIssuerNetworkServiceDatabaseFactory(pluginDatabaseSystem);
 
             try {
 
                 /*
                  * We create the new database
                  */
-                this.dataBase = communicationNetworkServiceDatabaseFactory.createDatabase(pluginId, AssetIssuerNetworkServiceDatabaseConstants.DATA_BASE_NAME);
+                this.dataBase = assetIssuerNetworkServiceDatabaseFactory.createDatabase(pluginId, AssetIssuerNetworkServiceDatabaseConstants.DATA_BASE_NAME);
 
             } catch (CantCreateDatabaseException cantOpenDatabaseException) {
 
                 /*
                  * The database cannot be created. I can not handle this situation.
                  */
-                //errorManager.reportUnexpectedPluginException(Plugins.BITDUBAI_DAP_ASSET_USER_ACTOR_NETWORK_SERVICE, UnexpectedPluginExceptionSeverity.DISABLES_SOME_FUNCTIONALITY_WITHIN_THIS_PLUGIN, cantOpenDatabaseException);
+                reportUnexpectedError(e);
                 throw new CantInitializeTemplateNetworkServiceDatabaseException(cantOpenDatabaseException.getLocalizedMessage());
 
             }
         }
     }
 
-    /**
-     * This method validate is all required resource are injected into
-     * the plugin root by the platform
-     *
-     * @throws CantStartPluginException
-     */
-    private void validateInjectedResources() throws CantStartPluginException {
-
-         /*
-         * If all resources are inject
-         */
-        if (wsCommunicationsCloudClientManager == null ||
-                pluginDatabaseSystem == null ||
-                errorManager == null ||
-                eventManager == null) {
-
-            StringBuffer contextBuffer = new StringBuffer();
-            contextBuffer.append("Plugin ID: " + pluginId);
-            contextBuffer.append(CantStartPluginException.CONTEXT_CONTENT_SEPARATOR);
-            contextBuffer.append("wsCommunicationsCloudClientManager: " + wsCommunicationsCloudClientManager);
-            contextBuffer.append(CantStartPluginException.CONTEXT_CONTENT_SEPARATOR);
-            contextBuffer.append("pluginDatabaseSystem: " + pluginDatabaseSystem);
-            contextBuffer.append(CantStartPluginException.CONTEXT_CONTENT_SEPARATOR);
-            contextBuffer.append("errorManager: " + errorManager);
-            contextBuffer.append(CantStartPluginException.CONTEXT_CONTENT_SEPARATOR);
-            contextBuffer.append("eventManager: " + eventManager);
-
-            String context = contextBuffer.toString();
-            String possibleCause = "No all required resource are injected";
-            CantStartPluginException pluginStartException = new CantStartPluginException(CantStartPluginException.DEFAULT_MESSAGE, null, context, possibleCause);
-
-            errorManager.reportUnexpectedPluginException(Plugins.BITDUBAI_DAP_ASSET_ISSUER_ACTOR_NETWORK_SERVICE, UnexpectedPluginExceptionSeverity.DISABLES_THIS_PLUGIN, pluginStartException);
-            throw pluginStartException;
-        }
-
-    }
-
-    @Override
-    public void pause() {
-        /*
-         * Pause
-         */
-        communicationNetworkServiceConnectionManager.pause();
-
-        /*
-         * Set the new status
-         */
-        this.serviceStatus = ServiceStatus.PAUSED;
-    }
-
-    @Override
-    public void resume() {
-        /*
-         * resume the managers
-         */
-        communicationNetworkServiceConnectionManager.resume();
-
-        /*
-         * Set the new status
-         */
-        this.serviceStatus = ServiceStatus.STARTED;
-    }
-
-    @Override
-    public void stop() {
-        //Clear all references of the event listeners registered with the event manager.
-        listenersAdded.clear();
-
-        /*
-         * Stop all connection on the managers
-         */
-        communicationNetworkServiceConnectionManager.closeAllConnection();
-
-        //set to not register
-        register = Boolean.FALSE;
-
-        /*
-         * Set the new status
-         */
-        this.serviceStatus = ServiceStatus.STOPPED;
-    }
-
-
-    /**
-     * Static method to get the logging level from any class under root.
-     *
-     * @param className
-     * @return
-     */
-    public static LogLevel getLogLevelByClass(String className) {
+    private void checkFailedDeliveryTime(String destinationPublicKey) {
         try {
-            /**
-             * sometimes the classname may be passed dinamically with an $moretext
-             * I need to ignore whats after this.
-             */
-            String[] correctedClass = className.split((Pattern.quote("$")));
-            return AssetIssuerActorNetworkServicePluginRoot.newLoggingLevel.get(correctedClass[0]);
+
+            List<ActorAssetNetworkServiceRecord> actorNetworkServiceRecordList = outgoingNotificationDao.getNotificationByDestinationPublicKey(destinationPublicKey);
+
+            //if I try to send more than 5 times I put it on hold
+            for (ActorAssetNetworkServiceRecord record : actorNetworkServiceRecordList) {
+
+                if (!record.getActorAssetProtocolState().getCode().equals(ActorAssetProtocolState.WAITING_RESPONSE.getCode())) {
+                    if (record.getSentCount() > 10) {
+                        //  if(record.getSentCount() > 20)
+                        //  {
+                        //reprocess at two hours
+                        //  reprocessTimer =  2 * 3600 * 1000;
+                        // }
+
+                        record.setActorAssetProtocolState(ActorAssetProtocolState.WAITING_RESPONSE);
+                        record.setSentCount(1);
+                        //update state and process again later
+
+                        outgoingNotificationDao.update(record);
+                    } else {
+                        record.setSentCount(record.getSentCount() + 1);
+                        outgoingNotificationDao.update(record);
+                    }
+                } else {
+                    //I verify the number of days I'm around trying to send if it exceeds three days I delete record
+                    long sentDate = record.getSentDate();
+                    long currentTime = System.currentTimeMillis();
+                    long dif = currentTime - sentDate;
+
+                    double dias = Math.floor(dif / (1000 * 60 * 60 * 24));
+
+                    if ((int) dias > 3) {
+                        //notify the user does not exist to actor asset user actor plugin
+                        record.changeDescriptor(AssetNotificationDescriptor.ACTOR_ASSET_NOT_FOUND);
+                        incomingNotificationsDao.createNotification(record);
+
+                        outgoingNotificationDao.delete(record.getId());
+                    }
+                }
+            }
         } catch (Exception e) {
-            /**
-             * If I couldn't get the correct loggin level, then I will set it to minimal.
-             */
-            return DEFAULT_LOG_LEVEL;
+            System.out.println("ACTOR ASSET ISSUER NS EXCEPCION VERIFICANDO WAIT MESSAGE");
+            e.printStackTrace();
         }
     }
 
-    @Override
-    public String getIdentityPublicKey() {
-        return this.identity.getPublicKey();
-    }
-
     /**
-     * This method initialize the communicationNetworkServiceConnectionManager.
-     * IMPORTANT: Call this method only in the CommunicationRegistrationProcessNetworkServiceAgent, when execute the registration process
-     * because at this moment, is create the platformComponentProfile for this component
+     * Mark the message as read
+     *
+     * @param fermatMessage
      */
-    public void initializeCommunicationNetworkServiceConnectionManager() {
-        this.communicationNetworkServiceConnectionManager = new CommunicationNetworkServiceConnectionManager(
-                getPlatformComponentProfilePluginRoot(),
-                identity,
-                wsCommunicationsCloudClientManager.getCommunicationsCloudClientConnection(),
-                dataBase,
-                errorManager,
-                eventManager);
-    }
-
-    private void initializeActorAssetIssuerAgent() {
+    public void markAsRead(FermatMessage fermatMessage) throws CantUpdateRecordDataBaseException {
         try {
-            assetIssuerActorNetworkServiceAgent = new AssetIssuerActorNetworkServiceAgent(
-                    this,
-                    wsCommunicationsCloudClientManager,
-                    communicationNetworkServiceConnectionManager,
-                    getPlatformComponentProfilePluginRoot(),
-                    errorManager,
-                    identity,
-                    dataBase);
-            assetIssuerActorNetworkServiceAgent.start();
-
-        } catch (CantStartAgentException e) {
-            errorManager.reportUnexpectedPluginException(Plugins.BITDUBAI_DAP_ASSET_ISSUER_ACTOR_NETWORK_SERVICE, UnexpectedPluginExceptionSeverity.DISABLES_THIS_PLUGIN, e);
+            ((FermatMessageCommunication) fermatMessage).setFermatMessagesStatus(FermatMessagesStatus.READ);
+            getCommunicationNetworkServiceConnectionManager().getIncomingMessageDao().update(fermatMessage);
+        } catch (Exception e) {
+            e.printStackTrace();
         }
     }
 
@@ -618,11 +667,8 @@ public class AssetIssuerActorNetworkServicePluginRoot extends AbstractNetworkSer
     public void registerActorAssetIssuer(ActorAssetIssuer actorAssetIssuerToRegister) throws CantRegisterActorAssetIssuerException {
 
         try {
-            final CommunicationsClientConnection communicationsClientConnection = wsCommunicationsCloudClientManager.getCommunicationsCloudClientConnection();
-            /*
-             * If register
-             */
-            if (true) {
+            if (isRegister()) {
+                final CommunicationsClientConnection communicationsClientConnection = wsCommunicationsCloudClientManager.getCommunicationsCloudClientConnection();
                 /*
                  * Construct the profile
                  */
@@ -644,39 +690,23 @@ public class AssetIssuerActorNetworkServicePluginRoot extends AbstractNetworkSer
                 /**
                  * I need to add this in a new thread other than the main android thread
                  */
+                if (!actorAssetIssuerPendingToRegistration.contains(platformComponentProfileAssetIssuer)) {
+                    actorAssetIssuerPendingToRegistration.add(platformComponentProfileAssetIssuer);
+                }
+
                 Thread thread = new Thread(new Runnable() {
                     @Override
                     public void run() {
                         try {
-                            communicationsClientConnection.registerComponentForCommunication(getNetworkServiceType(), platformComponentProfileAssetIssuer);
+                            communicationsClientConnection.registerComponentForCommunication(
+                                    getNetworkServiceProfile().getNetworkServiceType(), platformComponentProfileAssetIssuer);
+                            onComponentRegistered(platformComponentProfileAssetIssuer);
                         } catch (CantRegisterComponentException e) {
                             e.printStackTrace();
                         }
                     }
-                });
+                }, "ACTOR ASSET ISSUER REGISTER-ACTOR");
                 thread.start();
-
-
-            } else {
-                /*
-                 * Construct the profile
-                 */
-                Gson gson = new Gson();
-                JsonObject jsonObject = new JsonObject();
-                jsonObject.addProperty(DAP_IMG_ISSUER, Base64.encodeToString(actorAssetIssuerToRegister.getProfileImage(), Base64.DEFAULT));
-                String extraData = gson.toJson(jsonObject);
-
-                PlatformComponentProfile platformComponentProfileAssetIssuer = communicationsClientConnection.constructPlatformComponentProfileFactory(
-                        actorAssetIssuerToRegister.getActorPublicKey(),
-                        actorAssetIssuerToRegister.getName().toLowerCase().trim(),
-                        actorAssetIssuerToRegister.getName(),
-                        NetworkServiceType.UNDEFINED,
-                        PlatformComponentType.ACTOR_ASSET_ISSUER,
-                        extraData);
-                /*
-                 * Add to the list of pending to register
-                 */
-                actorAssetIssuerPendingToRegistration.add(platformComponentProfileAssetIssuer);
             }
         } catch (Exception e) {
             StringBuffer contextBuffer = new StringBuffer();
@@ -704,11 +734,8 @@ public class AssetIssuerActorNetworkServicePluginRoot extends AbstractNetworkSer
     public void updateActorAssetIssuer(ActorAssetIssuer actorAssetIssuerToRegister) throws CantRegisterActorAssetIssuerException {
 
         try {
-            final CommunicationsClientConnection communicationsClientConnection = wsCommunicationsCloudClientManager.getCommunicationsCloudClientConnection();
-            /*
-             * If register
-             */
-            if (true) {
+            if (isRegister()) {
+                final CommunicationsClientConnection communicationsClientConnection = wsCommunicationsCloudClientManager.getCommunicationsCloudClientConnection();
                 /*
                  * Construct the profile
                  */
@@ -729,36 +756,16 @@ public class AssetIssuerActorNetworkServicePluginRoot extends AbstractNetworkSer
                     @Override
                     public void run() {
                         try {
-                            communicationsClientConnection.updateRegisterActorProfile(getNetworkServiceType(), platformComponentProfileAssetIssuer);
+                            communicationsClientConnection.updateRegisterActorProfile(
+                                    getNetworkServiceProfile().getNetworkServiceType(), platformComponentProfileAssetIssuer);
+                            onComponentRegistered(platformComponentProfileAssetIssuer);
                         } catch (CantRegisterComponentException e) {
                             e.printStackTrace();
                         }
                     }
-                });
+                }, "ACTOR ASSET ISSUER UPDATE-ACTOR");
                 thread.start();
             }
-
-//            else {
-//                /*
-//                 * Construct the profile
-//                 */
-//                Gson gson = new Gson();
-//                JsonObject jsonObject = new JsonObject();
-//                jsonObject.addProperty(DAP_IMG_ISSUER, Base64.encodeToString(actorAssetIssuerToRegister.getProfileImage(), Base64.DEFAULT));
-//                String extraData = gson.toJson(jsonObject);
-//
-//                PlatformComponentProfile platformComponentProfileAssetIssuer = communicationsClientConnection.constructPlatformComponentProfileFactory(
-//                        actorAssetIssuerToRegister.getActorPublicKey(),
-//                        actorAssetIssuerToRegister.getName().toLowerCase().trim(),
-//                        actorAssetIssuerToRegister.getName(),
-//                        NetworkServiceType.UNDEFINED,
-//                        PlatformComponentType.ACTOR_ASSET_ISSUER,
-//                        extraData);
-//                /*
-//                 * Add to the list of pending to register
-//                 */
-//                actorAssetIssuerPendingToRegistration.add(platformComponentProfileAssetIssuer);
-//            }
         } catch (Exception e) {
             StringBuffer contextBuffer = new StringBuffer();
             contextBuffer.append("Plugin ID: " + pluginId);
@@ -781,253 +788,10 @@ public class AssetIssuerActorNetworkServicePluginRoot extends AbstractNetworkSer
         }
     }
 
-    public void handleNewMessages(final FermatMessage fermatMessage) {
-
-        try {
-
-            Gson gson = new Gson();
-            String jsonMessage = fermatMessage.getContent();
-            NetworkServiceMessage networkServiceDaoMessage = gson.fromJson(jsonMessage, NetworkServiceMessage.class);
-
-            switch (networkServiceDaoMessage.getMessageType()) {
-
-                case ACCEPT:
-                    NetworkServiceMessageAccept acceptMessage = gson.fromJson(jsonMessage, NetworkServiceMessageAccept.class);
-//                    receiveAcceptance(acceptMessage);
-                    break;
-
-                case DENY:
-                    NetworkServiceMessageDeny denyMessage = gson.fromJson(jsonMessage, NetworkServiceMessageDeny.class);
-//                    receiveDenial(denyMessage);
-                    break;
-
-                case REQUEST:
-                    // update the request to processing receive state with the given action.
-                    NetworkServiceMessageRequest requestMessage = gson.fromJson(jsonMessage, NetworkServiceMessageRequest.class);
-//                    receiveRequest(requestMessage);
-                    break;
-
-                default:
-                    throw new CantHandleDapNewMessagesException(
-                            "message type: " + networkServiceDaoMessage.getMessageType().name(),
-                            "Message type not handled."
-                    );
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-    }
-
-    @Override
-    public void handleNewSentMessageNotificationEvent(FermatMessage data) {
-
-    }
-
-//    @Override
-//    public void sendMessage(DAPMessage dapMessage) throws CantSendMessageException {
-//        switch (dapMessage.getMessageContent().messageType()) {
-//            case ASSET_APPROPRIATION:
-//                assetAppropriated(dapMessage);
-//                break;
-//            case EXTENDED_PUBLIC_KEY:
-//                requestPublicKeyExtended(dapMessage);
-//                break;
-//            default:
-//                throw new CantSendMessageException("This message is not registered or can't be handled by this network service..");
-//        }
-//    }
-
-    private void requestPublicKeyExtended(DAPMessage dapMessage) {
-        DAPActor actorRedeemPointSender = dapMessage.getActorSender();
-        DAPActor actorIssuerDestination = dapMessage.getActorReceiver();
-        try {
-            CommunicationNetworkServiceLocal communicationNetworkServiceLocal = communicationNetworkServiceConnectionManager.getNetworkServiceLocalInstance(actorIssuerDestination.getActorPublicKey());
-
-            if (true) {
-
-//                JsonObject jsonObject = new JsonObject();
-//                jsonObject.addProperty(DAP_IMG_ISSUER, Base64.encodeToString(dapMessage.getActorSender().getProfileImage(), Base64.DEFAULT));
-//                jsonObject.addProperty(DAP_IMG_ISSUER, Base64.encodeToString(dapMessage.getActorReceiver().getProfileImage(), Base64.DEFAULT));
-
-                String messageContentIntoJson = XMLParser.parseObject(dapMessage);
-
-                if (communicationNetworkServiceLocal != null) {
-
-                    //Send the message
-                    communicationNetworkServiceLocal.sendMessage(
-                            actorRedeemPointSender.getActorPublicKey(),
-                            actorIssuerDestination.getActorPublicKey(),
-                            messageContentIntoJson);
-
-                } else {
-
-                    /*
-                     * Created the message
-                     */
-                    FermatMessage fermatMessage = FermatMessageCommunicationFactory.constructFermatMessage(actorRedeemPointSender.getActorPublicKey(),//Sender
-                            actorIssuerDestination.getActorPublicKey(), //Receiver
-                            messageContentIntoJson,                //Message Content
-                            FermatMessageContentType.TEXT);//Type
-
-                    /*
-                     * Configure the correct status
-                     */
-                    ((FermatMessageCommunication) fermatMessage).setFermatMessagesStatus(FermatMessagesStatus.PENDING_TO_SEND);
-
-                    /*
-                     * Save to the data base table
-                     */
-                    OutgoingMessageDao outgoingMessageDao = communicationNetworkServiceConnectionManager.getOutgoingMessageDao();
-                    outgoingMessageDao.create(fermatMessage);
-
-                    /*
-                     * Create the sender basic profile
-                     */
-                    PlatformComponentProfile sender = wsCommunicationsCloudClientManager.getCommunicationsCloudClientConnection().
-                            constructBasicPlatformComponentProfileFactory(
-                                    actorRedeemPointSender.getActorPublicKey(),
-                                    NetworkServiceType.UNDEFINED,
-                                    PlatformComponentType.ACTOR_ASSET_REDEEM_POINT);
-
-                    /*
-                     * Create the receiver basic profile
-                     */
-                    PlatformComponentProfile receiver = wsCommunicationsCloudClientManager.getCommunicationsCloudClientConnection().
-                            constructBasicPlatformComponentProfileFactory(
-                                    actorIssuerDestination.getActorPublicKey(),
-                                    NetworkServiceType.UNDEFINED,
-                                    PlatformComponentType.ACTOR_ASSET_ISSUER);
-
-                    /*
-                     * Ask the client to connect
-                     */
-                    communicationNetworkServiceConnectionManager.connectTo(sender, getPlatformComponentProfilePluginRoot(), receiver);
-
-                }
-            } else {
-
-                StringBuffer contextBuffer = new StringBuffer();
-                contextBuffer.append("Plugin ID: " + pluginId);
-                contextBuffer.append(CantStartPluginException.CONTEXT_CONTENT_SEPARATOR);
-                contextBuffer.append("wsCommunicationsCloudClientManager: " + wsCommunicationsCloudClientManager);
-                contextBuffer.append(CantStartPluginException.CONTEXT_CONTENT_SEPARATOR);
-                contextBuffer.append("pluginDatabaseSystem: " + pluginDatabaseSystem);
-                contextBuffer.append(CantStartPluginException.CONTEXT_CONTENT_SEPARATOR);
-                contextBuffer.append("errorManager: " + errorManager);
-                contextBuffer.append(CantStartPluginException.CONTEXT_CONTENT_SEPARATOR);
-                contextBuffer.append("eventManager: " + eventManager);
-
-                String context = contextBuffer.toString();
-                String possibleCause = "Asset Issuer Actor Network Service Not Registered";
-
-            }
-        } catch (Exception e) {
-
-            StringBuffer contextBuffer = new StringBuffer();
-            contextBuffer.append("Plugin ID: " + pluginId);
-            contextBuffer.append(CantStartPluginException.CONTEXT_CONTENT_SEPARATOR);
-            contextBuffer.append("wsCommunicationsCloudClientManager: " + wsCommunicationsCloudClientManager);
-            contextBuffer.append(CantStartPluginException.CONTEXT_CONTENT_SEPARATOR);
-            contextBuffer.append("pluginDatabaseSystem: " + pluginDatabaseSystem);
-            contextBuffer.append(CantStartPluginException.CONTEXT_CONTENT_SEPARATOR);
-            contextBuffer.append("errorManager: " + errorManager);
-            contextBuffer.append(CantStartPluginException.CONTEXT_CONTENT_SEPARATOR);
-            contextBuffer.append("eventManager: " + eventManager);
-
-            String context = contextBuffer.toString();
-            String possibleCause = "Cant Request Public Key Extended";
-
-//            CantRequestCryptoAddressException pluginStartException = new CantRequestCryptoAddressException(CantStartPluginException.DEFAULT_MESSAGE, null, context, possibleCause);
-//            errorManager.reportUnexpectedPluginException(Plugins.BITDUBAI_DAP_ASSET_ISSUER_ACTOR_NETWORK_SERVICE, UnexpectedPluginExceptionSeverity.DISABLES_THIS_PLUGIN, pluginStartException);
-
-//            throw pluginStartException;
-        }
-    }
-
-    private void assetAppropriated(DAPMessage dapMessage) throws CantSendMessageException {
-
-        String context = "Message: " + dapMessage;
-
-        if (!(dapMessage.getActorSender() instanceof ActorAssetUser) || !(dapMessage.getActorReceiver() instanceof ActorAssetIssuer)) {
-            throw new CantSendMessageException("One or both actors are from an incorrent type, please check", null, context);
-        }
-
-        DAPActor actorAssetUserSender = dapMessage.getActorSender();
-        DAPActor actorAssetIssuerReceiver = dapMessage.getActorReceiver();
-        try {
-            CommunicationNetworkServiceLocal communicationNetworkServiceLocal = communicationNetworkServiceConnectionManager.getNetworkServiceLocalInstance(actorAssetIssuerReceiver.getActorPublicKey());
-
-            if (true) {
-                String messageContentIntoJson = XMLParser.parseObject(dapMessage);
-
-                if (communicationNetworkServiceLocal != null) {
-
-                    //Send the message
-                    communicationNetworkServiceLocal.sendMessage(
-                            actorAssetUserSender.getActorPublicKey(),
-                            actorAssetIssuerReceiver.getActorPublicKey(),
-                            messageContentIntoJson);
-
-                } else {
-                    /*
-                     * Created the message
-                     */
-                    FermatMessage fermatMessage = FermatMessageCommunicationFactory.constructFermatMessage(
-                            actorAssetUserSender.getActorPublicKey(),//Sender
-                            actorAssetIssuerReceiver.getActorPublicKey(), //Receiver
-                            messageContentIntoJson,                //Message Content
-                            FermatMessageContentType.TEXT);//Type
-
-                    /*
-                     * Configure the correct status
-                     */
-                    ((FermatMessageCommunication) fermatMessage).setFermatMessagesStatus(FermatMessagesStatus.PENDING_TO_SEND);
-
-                    /*
-                     * Save to the data base table
-                     */
-                    OutgoingMessageDao outgoingMessageDao = communicationNetworkServiceConnectionManager.getOutgoingMessageDao();
-                    outgoingMessageDao.create(fermatMessage);
-
-                    /*
-                     * Create the sender basic profile
-                     */
-                    PlatformComponentProfile sender = wsCommunicationsCloudClientManager.getCommunicationsCloudClientConnection().
-                            constructBasicPlatformComponentProfileFactory(
-                                    actorAssetUserSender.getActorPublicKey(),
-                                    NetworkServiceType.UNDEFINED,
-                                    PlatformComponentType.ACTOR_ASSET_USER);
-
-                    /*
-                     * Create the receiver basic profile
-                     */
-                    PlatformComponentProfile receiver = wsCommunicationsCloudClientManager.getCommunicationsCloudClientConnection().
-                            constructBasicPlatformComponentProfileFactory(
-                                    actorAssetIssuerReceiver.getActorPublicKey(),
-                                    NetworkServiceType.UNDEFINED,
-                                    PlatformComponentType.ACTOR_ASSET_ISSUER);
-
-                    /*
-                     * Ask the client to connect
-                     */
-                    communicationNetworkServiceConnectionManager.connectTo(sender, getPlatformComponentProfilePluginRoot(), receiver);
-
-                }
-            } else {
-                throw new CantSendMessageException("The actor asset issuer network service is not registered.");
-            }
-        } catch (Exception e) {
-            throw new CantSendMessageException("Somethind bad happen while trying to send the asset appropriated message", e, context);
-        }
-    }
-
     @Override
     public List<ActorAssetIssuer> getListActorAssetIssuerRegistered() throws CantRequestListActorAssetIssuerRegisteredException {
 
         try {
-
-//            if (true) {
-
             if (actorAssetIssuerRegisteredList != null && !actorAssetIssuerRegisteredList.isEmpty()) {
                 actorAssetIssuerRegisteredList.clear();
             }
@@ -1054,7 +818,8 @@ public class AssetIssuerActorNetworkServicePluginRoot extends AbstractNetworkSer
                 for (PlatformComponentProfile platformComponentProfile : platformComponentProfileRegisteredListRemote) {
 
                     String profileImage = "";
-                    if (!platformComponentProfile.getExtraData().equals("")) {
+                    byte[] imageByte = null;
+                    if (Validate.isValidString(platformComponentProfile.getExtraData())) {
                         try {
                             JsonParser jParser = new JsonParser();
                             JsonObject jsonObject = jParser.parse(platformComponentProfile.getExtraData()).getAsJsonObject();
@@ -1063,9 +828,9 @@ public class AssetIssuerActorNetworkServicePluginRoot extends AbstractNetworkSer
                         } catch (Exception e) {
                             profileImage = platformComponentProfile.getExtraData();
                         }
+                        imageByte = Base64.decode(profileImage, Base64.DEFAULT);
                     }
 
-                    byte[] imageByte = Base64.decode(profileImage, Base64.DEFAULT);
 
                     ActorAssetIssuer actorAssetIssuerNew = new AssetIssuerActorRecord(
                             platformComponentProfile.getIdentityPublicKey(),
@@ -1097,7 +862,7 @@ public class AssetIssuerActorNetworkServicePluginRoot extends AbstractNetworkSer
 
             CantRequestListActorAssetIssuerRegisteredException pluginStartException = new CantRequestListActorAssetIssuerRegisteredException(CantRequestListActorAssetIssuerRegisteredException.DEFAULT_MESSAGE, null, context, possibleCause);
 
-            //errorManager.reportUnexpectedPluginException(Plugins.BITDUBAI_DAP_ASSET_USER_ACTOR_NETWORK_SERVICE, UnexpectedPluginExceptionSeverity.DISABLES_THIS_PLUGIN, pluginStartException);
+            errorManager.reportUnexpectedPluginException(Plugins.BITDUBAI_DAP_ASSET_ISSUER_ACTOR_NETWORK_SERVICE, UnexpectedPluginExceptionSeverity.DISABLES_THIS_PLUGIN, pluginStartException);
 
             throw pluginStartException;
         }
@@ -1105,152 +870,363 @@ public class AssetIssuerActorNetworkServicePluginRoot extends AbstractNetworkSer
         return actorAssetIssuerRegisteredList;
     }
 
-    @Override
-    public void askConnectionActorAsset(String actorAssetLoggedInPublicKey, String actorAssetLoggedName, Actors senderType, String actorAssetToAddPublicKey, String actorAssetToAddName, Actors destinationType, byte[] profileImage, BlockchainNetworkType blockchainNetworkType) throws CantAskConnectionActorAssetException {
 
+    @Override
+    public void askConnectionActorAsset(String actorAssetLoggedInPublicKey,
+                                        String actorAssetLoggedName,
+                                        Actors senderType,
+                                        String actorAssetToAddPublicKey,
+                                        String actorAssetToAddName,
+                                        Actors destinationType,
+                                        byte[] profileImage,
+                                        BlockchainNetworkType blockchainNetworkType) throws CantAskConnectionActorAssetException {
+
+        try {
+            UUID newNotificationID = UUID.randomUUID();
+            AssetNotificationDescriptor assetNotificationDescriptor = AssetNotificationDescriptor.ASKFORCONNECTION;
+            long currentTime = System.currentTimeMillis();
+            ActorAssetProtocolState actorAssetProtocolState = ActorAssetProtocolState.PROCESSING_SEND;
+
+            final ActorAssetNetworkServiceRecord assetIssuerNetworkServiceRecord = outgoingNotificationDao.createNotification(
+                    newNotificationID,
+                    actorAssetLoggedInPublicKey,
+                    senderType,
+                    actorAssetToAddPublicKey,
+                    actorAssetLoggedName,
+//                    intraUserToAddPhrase,
+                    profileImage,
+                    destinationType,
+                    assetNotificationDescriptor,
+                    currentTime,
+                    actorAssetProtocolState,
+                    false,
+                    1,
+                    blockchainNetworkType,
+                    null,
+                    null
+            );
+
+            executorService.submit(new Runnable() {
+                @Override
+                public void run() {
+                    try {
+                        sendNewMessage(
+                                getProfileSenderToRequestConnection(
+                                        assetIssuerNetworkServiceRecord.getActorSenderPublicKey(),
+                                        NetworkServiceType.UNDEFINED,
+                                        platformComponentTypeSelectorByActorType(assetIssuerNetworkServiceRecord.getActorSenderType())
+                                ),
+                                getProfileDestinationToRequestConnection(
+                                        assetIssuerNetworkServiceRecord.getActorDestinationPublicKey(),
+                                        NetworkServiceType.UNDEFINED,
+                                        platformComponentTypeSelectorByActorType(assetIssuerNetworkServiceRecord.getActorDestinationType())
+                                ),
+                                assetIssuerNetworkServiceRecord.toJson());
+                    } catch (com.bitdubai.fermat_p2p_api.layer.all_definition.communication.network_services.exceptions.CantSendMessageException | InvalidParameterException e) {
+                        e.printStackTrace();
+                    }
+                }
+            });
+            // Sending message to the destination
+        } catch (final CantCreateActorAssetNotificationException e) {
+            reportUnexpectedError(e);
+            throw new CantAskConnectionActorAssetException(e, "actor asset issuer network service", "database corrupted");
+        } catch (final Exception e) {
+            reportUnexpectedError(e);
+            throw new CantAskConnectionActorAssetException(e, "actor asset issuer network service", "Unhandled error.");
+        }
     }
 
     @Override
-    public void acceptConnectionActorAsset(String actorAssetLoggedInPublicKey, ActorAssetIssuer ActorAssetAccepted) throws CantAcceptConnectionActorAssetException {
+    public void acceptConnectionActorAsset(String actorAssetLoggedInPublicKey, ActorAssetIssuer actorAssetAccepted)
+            throws CantAcceptConnectionActorAssetException {
 
+        try {
+            ActorAssetNetworkServiceRecord assetIssuerNetworkServiceRecord = incomingNotificationsDao.
+                    changeActorAssetNotificationDescriptor(
+                            actorAssetAccepted.getActorPublicKey(),
+                            AssetNotificationDescriptor.ACCEPTED,
+                            ActorAssetProtocolState.DONE);
+//TODO Evaluar diferencias en ActorAssetProtocolState.DONE y ActorAssetProtocolState.PEDNING_ACTION para conocer diferencias
+
+            Actors actorSwap = assetIssuerNetworkServiceRecord.getActorSenderType();
+
+            assetIssuerNetworkServiceRecord.setActorSenderPublicKey(actorAssetLoggedInPublicKey);
+            assetIssuerNetworkServiceRecord.setActorSenderType(assetIssuerNetworkServiceRecord.getActorDestinationType());
+
+            assetIssuerNetworkServiceRecord.setActorDestinationPublicKey(actorAssetAccepted.getActorPublicKey());
+            assetIssuerNetworkServiceRecord.setActorDestinationType(actorSwap);
+
+            assetIssuerNetworkServiceRecord.setActorSenderAlias(actorAssetAccepted.getName());
+
+            assetIssuerNetworkServiceRecord.changeDescriptor(AssetNotificationDescriptor.ACCEPTED);
+
+            assetIssuerNetworkServiceRecord.changeState(ActorAssetProtocolState.PROCESSING_SEND);
+
+            final ActorAssetNetworkServiceRecord messageToSend = outgoingNotificationDao.createNotification(
+                    UUID.randomUUID(),
+                    assetIssuerNetworkServiceRecord.getActorSenderPublicKey(),
+                    assetIssuerNetworkServiceRecord.getActorSenderType(),
+                    assetIssuerNetworkServiceRecord.getActorDestinationPublicKey(),
+                    assetIssuerNetworkServiceRecord.getActorSenderAlias(),
+                    assetIssuerNetworkServiceRecord.getActorSenderProfileImage(),
+                    assetIssuerNetworkServiceRecord.getActorDestinationType(),
+                    assetIssuerNetworkServiceRecord.getAssetNotificationDescriptor(),
+                    System.currentTimeMillis(),
+                    assetIssuerNetworkServiceRecord.getActorAssetProtocolState(),
+                    false,
+                    1,
+                    assetIssuerNetworkServiceRecord.getBlockchainNetworkType(),
+                    assetIssuerNetworkServiceRecord.getResponseToNotificationId(),
+                    null
+            );
+
+            executorService.submit(new Runnable() {
+                @Override
+                public void run() {
+                    try {
+                        // Sending message to the destination
+                        sendNewMessage(
+                                getProfileSenderToRequestConnection(
+                                        messageToSend.getActorSenderPublicKey(),
+                                        NetworkServiceType.UNDEFINED,
+                                        platformComponentTypeSelectorByActorType(messageToSend.getActorSenderType())
+                                ),
+                                getProfileDestinationToRequestConnection(
+                                        messageToSend.getActorDestinationPublicKey(),
+                                        NetworkServiceType.UNDEFINED,
+                                        platformComponentTypeSelectorByActorType(messageToSend.getActorDestinationType())
+                                ),
+                                messageToSend.toJson());
+                    } catch (com.bitdubai.fermat_p2p_api.layer.all_definition.communication.network_services.exceptions.CantSendMessageException | InvalidParameterException e) {
+                        e.printStackTrace();
+                    }
+                }
+            });
+        } catch (Exception e) {
+            reportUnexpectedError(e);
+            throw new CantAcceptConnectionActorAssetException("ERROR ACTOR ASSET ISSUER NS WHEN ACCEPTING CONNECTION", e, "", "Generic Exception");
+        }
     }
 
     @Override
-    public void denyConnectionActorAsset(String actorAssetLoggedInPublicKey, ActorAssetIssuer actorAssetReject) throws CantDenyConnectionActorAssetException {
+    public void denyConnectionActorAsset(String actorAssetLoggedInPublicKey, ActorAssetIssuer actorAssetReject)
+            throws CantDenyConnectionActorAssetException {
 
+        try {
+            final ActorAssetNetworkServiceRecord actorNetworkServiceRecord = incomingNotificationsDao.
+                    changeActorAssetNotificationDescriptor(
+                            actorAssetReject.getActorPublicKey(),
+                            AssetNotificationDescriptor.DENIED,
+                            ActorAssetProtocolState.DONE);
+
+            Actors actorSwap = actorNetworkServiceRecord.getActorSenderType();
+
+            actorNetworkServiceRecord.setActorDestinationPublicKey(actorAssetReject.getActorPublicKey());
+            actorNetworkServiceRecord.setActorSenderType(actorNetworkServiceRecord.getActorDestinationType());
+
+            actorNetworkServiceRecord.setActorSenderPublicKey(actorAssetLoggedInPublicKey);
+            actorNetworkServiceRecord.setActorDestinationType(actorSwap);
+
+
+            actorNetworkServiceRecord.changeDescriptor(AssetNotificationDescriptor.DENIED);
+
+            actorNetworkServiceRecord.changeState(ActorAssetProtocolState.PROCESSING_SEND);
+
+            outgoingNotificationDao.createNotification(actorNetworkServiceRecord);
+
+            executorService.submit(new Runnable() {
+                @Override
+                public void run() {
+                    // Sending message to the destination
+                    try {
+                        sendNewMessage(
+                                getProfileSenderToRequestConnection(
+                                        actorNetworkServiceRecord.getActorSenderPublicKey(),
+                                        NetworkServiceType.UNDEFINED,
+                                        platformComponentTypeSelectorByActorType(actorNetworkServiceRecord.getActorSenderType())
+                                ),
+                                getProfileDestinationToRequestConnection(
+                                        actorNetworkServiceRecord.getActorDestinationPublicKey(),
+                                        NetworkServiceType.UNDEFINED,
+                                        platformComponentTypeSelectorByActorType(actorNetworkServiceRecord.getActorDestinationType())
+                                ),
+                                actorNetworkServiceRecord.toJson());
+                    } catch (com.bitdubai.fermat_p2p_api.layer.all_definition.communication.network_services.exceptions.CantSendMessageException | InvalidParameterException e) {
+                        e.printStackTrace();
+                    }
+                }
+            });
+        } catch (Exception e) {
+            reportUnexpectedError(e);
+            throw new CantDenyConnectionActorAssetException("ERROR DENY CONNECTION TO ACTOR ASSET ISSUER", e, "", "Generic Exception");
+        }
     }
 
     @Override
-    public void disconnectConnectionActorAsset(String actorAssetLoggedInPublicKey, String actorAssetToDisconnectPublicKey) throws CantDisconnectConnectionActorAssetException {
+    public void disconnectConnectionActorAsset(String actorAssetLoggedInPublicKey, String actorAssetToDisconnectPublicKey)
+            throws CantDisconnectConnectionActorAssetException {
 
+        try {
+            ActorAssetNetworkServiceRecord assetIssuerNetworkServiceRecord = incomingNotificationsDao.
+                    changeActorAssetNotificationDescriptor(
+                            actorAssetToDisconnectPublicKey,
+                            AssetNotificationDescriptor.DISCONNECTED,
+                            ActorAssetProtocolState.DONE);
+
+            Actors actorSwap = assetIssuerNetworkServiceRecord.getActorSenderType();
+
+            assetIssuerNetworkServiceRecord.setActorSenderPublicKey(actorAssetLoggedInPublicKey);
+            assetIssuerNetworkServiceRecord.setActorSenderType(assetIssuerNetworkServiceRecord.getActorDestinationType());
+
+            assetIssuerNetworkServiceRecord.setActorDestinationPublicKey(actorAssetToDisconnectPublicKey);
+            assetIssuerNetworkServiceRecord.setActorDestinationType(actorSwap);
+
+            assetIssuerNetworkServiceRecord.setActorSenderAlias(null);
+
+            assetIssuerNetworkServiceRecord.changeDescriptor(AssetNotificationDescriptor.DISCONNECTED);
+
+            assetIssuerNetworkServiceRecord.changeState(ActorAssetProtocolState.PROCESSING_SEND);
+
+            //make message to actor
+//            UUID newNotificationID = UUID.randomUUID();
+//            AssetNotificationDescriptor assetNotificationDescriptor = AssetNotificationDescriptor.DISCONNECTED;
+//            long currentTime = System.currentTimeMillis();
+//            ActorAssetProtocolState actorAssetProtocolState = ActorAssetProtocolState.PROCESSING_SEND;
+
+            final ActorAssetNetworkServiceRecord actorNetworkServiceRecord = outgoingNotificationDao.createNotification(
+                    UUID.randomUUID(),
+                    assetIssuerNetworkServiceRecord.getActorSenderPublicKey(),
+                    assetIssuerNetworkServiceRecord.getActorSenderType(),
+                    assetIssuerNetworkServiceRecord.getActorDestinationPublicKey(),
+                    assetIssuerNetworkServiceRecord.getActorSenderAlias(),
+                    assetIssuerNetworkServiceRecord.getActorSenderProfileImage(),
+                    assetIssuerNetworkServiceRecord.getActorDestinationType(),
+                    assetIssuerNetworkServiceRecord.getAssetNotificationDescriptor(),
+                    System.currentTimeMillis(),
+                    assetIssuerNetworkServiceRecord.getActorAssetProtocolState(),
+                    false,
+                    1,
+                    assetIssuerNetworkServiceRecord.getBlockchainNetworkType(),
+                    null,
+                    null
+            );
+
+            executorService.submit(new Runnable() {
+                @Override
+                public void run() {
+                    // Sending message to the destination
+                    try {
+                        sendNewMessage(
+                                getProfileSenderToRequestConnection(
+                                        actorNetworkServiceRecord.getActorSenderPublicKey(),
+                                        NetworkServiceType.UNDEFINED,
+                                        platformComponentTypeSelectorByActorType(actorNetworkServiceRecord.getActorSenderType())
+                                ),
+                                getProfileDestinationToRequestConnection(
+                                        actorNetworkServiceRecord.getActorDestinationPublicKey(),
+                                        NetworkServiceType.UNDEFINED,
+                                        platformComponentTypeSelectorByActorType(actorNetworkServiceRecord.getActorDestinationType())
+                                ),
+                                actorNetworkServiceRecord.toJson());
+                    } catch (com.bitdubai.fermat_p2p_api.layer.all_definition.communication.network_services.exceptions.CantSendMessageException | InvalidParameterException e) {
+                        e.printStackTrace();
+                    }
+                }
+            });
+        } catch (Exception e) {
+            reportUnexpectedError(e);
+            throw new CantDisconnectConnectionActorAssetException("ERROR DISCONNECTING ACTOR ASSET ISSUER ", e, "", "Generic Exception");
+        }
     }
 
     @Override
-    public void cancelConnectionActorAsset(String actorAssetLoggedInPublicKey, ActorAssetIssuer actorAssetToCancel) throws CantCancelConnectionActorAssetException {
+    public void cancelConnectionActorAsset(String actorAssetLoggedInPublicKey, ActorAssetIssuer actorAssetToCancel)
+            throws CantCancelConnectionActorAssetException {
 
+        try {
+            final ActorAssetNetworkServiceRecord assetIssuerNetworkServiceRecord = incomingNotificationsDao.
+                    changeActorAssetNotificationDescriptor(
+                            actorAssetToCancel.getActorPublicKey(),
+                            AssetNotificationDescriptor.CANCEL,
+                            ActorAssetProtocolState.DONE);
+
+            Actors actorSwap = assetIssuerNetworkServiceRecord.getActorSenderType();
+
+            assetIssuerNetworkServiceRecord.setActorSenderPublicKey(actorAssetLoggedInPublicKey);
+            assetIssuerNetworkServiceRecord.setActorSenderType(assetIssuerNetworkServiceRecord.getActorDestinationType());
+
+            assetIssuerNetworkServiceRecord.setActorDestinationPublicKey(actorAssetToCancel.getActorPublicKey());
+            assetIssuerNetworkServiceRecord.setActorDestinationType(actorSwap);
+
+//            assetIssuerNetworkServiceRecord.setActorSenderAlias(actorAssetAccepted.getName());
+
+            assetIssuerNetworkServiceRecord.changeDescriptor(AssetNotificationDescriptor.CANCEL);
+
+            assetIssuerNetworkServiceRecord.changeState(ActorAssetProtocolState.PROCESSING_SEND);
+
+            outgoingNotificationDao.createNotification(assetIssuerNetworkServiceRecord);
+
+            executorService.submit(new Runnable() {
+                @Override
+                public void run() {
+                    // Sending message to the destination
+                    try {
+                        sendNewMessage(
+                                getProfileSenderToRequestConnection(
+                                        assetIssuerNetworkServiceRecord.getActorSenderPublicKey(),
+                                        NetworkServiceType.UNDEFINED,
+                                        platformComponentTypeSelectorByActorType(assetIssuerNetworkServiceRecord.getActorSenderType())
+                                ),
+                                getProfileDestinationToRequestConnection(
+                                        assetIssuerNetworkServiceRecord.getActorDestinationPublicKey(),
+                                        NetworkServiceType.UNDEFINED,
+                                        platformComponentTypeSelectorByActorType(assetIssuerNetworkServiceRecord.getActorDestinationType())
+                                ),
+                                assetIssuerNetworkServiceRecord.toJson());
+                    } catch (com.bitdubai.fermat_p2p_api.layer.all_definition.communication.network_services.exceptions.CantSendMessageException | InvalidParameterException e) {
+                        e.printStackTrace();
+                    }
+                }
+            });
+        } catch (Exception e) {
+            reportUnexpectedError(e);
+            throw new CantCancelConnectionActorAssetException("ERROR CANCEL CONNECTION TO ACTOR ASSET ISSUER ", e, "", "Generic Exception");
+        }
     }
 
     @Override
     public List<ActorNotification> getPendingNotifications() throws CantGetActorAssetNotificationException {
-        return null;
+        try {
+            if(incomingNotificationsDao == null)
+                incomingNotificationsDao = new IncomingNotificationDao(dataBase, pluginFileSystem, pluginId);
+            return incomingNotificationsDao.listUnreadNotifications();
+
+
+        } catch (CantGetActorAssetNotificationException e) {
+            reportUnexpectedError(e);
+            throw new CantGetActorAssetNotificationException(e, "ACTOR ASSET ISSUER network service", "database corrupted");
+        } catch (Exception e) {
+            reportUnexpectedError(e);
+            throw new CantGetActorAssetNotificationException(e, "ACTOR ASSET ISSUER network service", "Unhandled error.");
+        }
     }
 
     @Override
     public void confirmActorAssetNotification(UUID notificationID) throws CantConfirmActorAssetNotificationException {
-
-    }
-
-    @Override
-    public void buildSendMessage(ActorNotification actorNotification) {
-
-    }
-
-    @Override
-    public void responseExtended(ActorNotification actorNotification, ExtendedPublicKey extendedPublicKey) {
-
-    }
-
-    @Override
-    public List<PlatformComponentProfile> getRemoteNetworkServicesRegisteredList() {
-        return remoteNetworkServicesRegisteredList;
-    }
-
-    @Override
-    public void requestRemoteNetworkServicesRegisteredList(DiscoveryQueryParameters discoveryQueryParameters) {
-        System.out.println(" AssetIssuerActorNetworkServicePluginRoot - requestRemoteNetworkServicesRegisteredList");
-
-        /*
-         * Request the list of component registers
-         */
         try {
-            wsCommunicationsCloudClientManager.getCommunicationsCloudClientConnection().requestListComponentRegistered(getPlatformComponentProfilePluginRoot(), discoveryQueryParameters);
-        } catch (CantRequestListException e) {
-            e.printStackTrace();
+            incomingNotificationsDao.markNotificationAsRead(notificationID);
+        } catch (final Exception e) {
+            reportUnexpectedError(e);
+            throw new CantConfirmActorAssetNotificationException(e, "notificationID: " + notificationID, "Unhandled error.");
         }
     }
 
     @Override
-    public NetworkServiceConnectionManager getNetworkServiceConnectionManager() {
-        return communicationNetworkServiceConnectionManager;
-    }
-
-    @Override
-    public DiscoveryQueryParameters constructDiscoveryQueryParamsFactory(PlatformComponentType platformComponentType, NetworkServiceType networkServiceType, String alias, String identityPublicKey, Location location, Double distance, String name, String extraData, Integer firstRecord, Integer numRegister, PlatformComponentType fromOtherPlatformComponentType, NetworkServiceType fromOtherNetworkServiceType) {
-        return wsCommunicationsCloudClientManager.getCommunicationsCloudClientConnection().constructDiscoveryQueryParamsFactory(platformComponentType,
-                networkServiceType, alias, identityPublicKey, location, distance, name, extraData, firstRecord, numRegister, fromOtherPlatformComponentType, fromOtherNetworkServiceType);
-
-    }
-
-    @Override
-    public void handleCompleteComponentRegistrationNotificationEvent(PlatformComponentProfile platformComponentProfileRegistered) {
-        if (platformComponentProfileRegistered.getPlatformComponentType() == PlatformComponentType.COMMUNICATION_CLOUD_CLIENT && this.register) {
-            if (communicationRegistrationProcessNetworkServiceAgent.isAlive()) {
-                communicationRegistrationProcessNetworkServiceAgent.interrupt();
-                communicationRegistrationProcessNetworkServiceAgent = null;
-            }
-
-                /*
-                 * Construct my profile and register me
-                 */
-            PlatformComponentProfile platformComponentProfileToReconnect = wsCommunicationsCloudClientManager.getCommunicationsCloudClientConnection().constructPlatformComponentProfileFactory(this.getIdentityPublicKey(),
-
-                    this.getAlias().toLowerCase(),
-                    this.getName(),
-                    this.getNetworkServiceType(),
-                    this.getPlatformComponentType(),
-                    this.getExtraData());
-
-            try {
-                    /*
-                     * Register me
-                     */
-                wsCommunicationsCloudClientManager.getCommunicationsCloudClientConnection().registerComponentForCommunication(this.getNetworkServiceType(), platformComponentProfileToReconnect);
-
-            } catch (CantRegisterComponentException e) {
-                e.printStackTrace();
-            }
-
-        }
+    public void onComponentRegistered(PlatformComponentProfile platformComponentProfileRegistered) {
 
         if (platformComponentProfileRegistered.getPlatformComponentType() == PlatformComponentType.ACTOR_ASSET_ISSUER) {
-            System.out.print("ACTOR ASSET ISSUER REGISTRADO!! -----------------------\n" +
-                    "-----------------------\n ISSUER: " + platformComponentProfileRegistered.getAlias());
-        }
-
-        /*
-         * If the component registered have my profile and my identity public key
-         */
-        if (platformComponentProfileRegistered.getPlatformComponentType() == PlatformComponentType.NETWORK_SERVICE &&
-                platformComponentProfileRegistered.getNetworkServiceType() == NetworkServiceType.ASSET_ISSUER_ACTOR &&
-                platformComponentProfileRegistered.getIdentityPublicKey().equals(identity.getPublicKey())) {
-
-            /*
-             * Mark as register
-             */
-            this.register = Boolean.TRUE;
-
-            initializeActorAssetIssuerAgent();
-
-            /*
-             * If exist actor asset Issuer pending to registration
-             */
-            if (actorAssetIssuerPendingToRegistration != null && !actorAssetIssuerPendingToRegistration.isEmpty()) {
-
-                CommunicationsClientConnection communicationsClientConnection = wsCommunicationsCloudClientManager.getCommunicationsCloudClientConnection();
-
-                for (PlatformComponentProfile platformComponentProfileAssetIssuer : actorAssetIssuerPendingToRegistration) {
-                     /*
-                     * ask to the communication cloud client to register
-                     */
-                    try {
-                        communicationsClientConnection.registerComponentForCommunication(
-                                this.getNetworkServiceType(),
-                                platformComponentProfileAssetIssuer);
-                    } catch (CantRegisterComponentException e) {
-                        e.printStackTrace();
-                    }
-                }
-            }
+            System.out.print("ACTOR ASSET ISSUER REGISTRADO: " + platformComponentProfileRegistered.getAlias());
         }
 
         /*
@@ -1264,42 +1240,11 @@ public class AssetIssuerActorNetworkServicePluginRoot extends AbstractNetworkSer
             ActorAssetIssuer actorAssetIssuerNewRegistered = new AssetIssuerActorRecord(
                     platformComponentProfileRegistered.getIdentityPublicKey(),
                     platformComponentProfileRegistered.getName(),
-                    convertoByteArrayfromString(platformComponentProfileRegistered.getExtraData()),
+                    new byte[0],
+//                    convertoByteArrayfromString(platformComponentProfileRegistered.getExtraData()),
                     loca);
 
-            if (communicationNetworkServiceConnectionManager == null) {
-
-                this.communicationNetworkServiceConnectionManager = new CommunicationNetworkServiceConnectionManager(
-                        getPlatformComponentProfilePluginRoot(),
-                        identity,
-                        wsCommunicationsCloudClientManager.getCommunicationsCloudClientConnection(),
-                        dataBase,
-                        errorManager,
-                        eventManager);
-            }
-//            if (createactorAgentnum == 0) {
-//
-//                assetIssuerActorNetworkServiceAgent = new AssetIssuerActorNetworkServiceAgent(
-//                        this,
-//                        wsCommunicationsCloudClientManager,
-//                        communicationNetworkServiceConnectionManager,
-//                        getPlatformComponentProfilePluginRoot(),
-//                        errorManager,
-//                        identity,
-//                        dataBase);
-//
-//                // start main threads
-//                try {
-//                    assetIssuerActorNetworkServiceAgent.start();
-//                } catch (CantStartAgentException e) {
-//                    e.printStackTrace();
-//                }
-//
-//                createactorAgentnum = 1;
-//
-//            }
-
-            System.out.println("Actor Asset Issuer REGISTRADO en A.N.S - Enviando Evento de Notificacion");
+            System.out.println("ACTOR ASSET ISSUER REGISTRADO en A.N.S - Enviando Evento de Notificacion");
 
             FermatEvent event = eventManager.getNewEvent(EventType.COMPLETE_ASSET_ISSUER_REGISTRATION_NOTIFICATION);
             event.setSource(EventSource.ACTOR_ASSET_ISSUER);
@@ -1310,220 +1255,163 @@ public class AssetIssuerActorNetworkServicePluginRoot extends AbstractNetworkSer
         }
     }
 
-    @Override
-    public void handleFailureComponentRegistrationNotificationEvent(PlatformComponentProfile networkServiceApplicant, PlatformComponentProfile remoteParticipant) {
-        System.out.println(" CommunicationNetworkServiceConnectionManager - Starting method handleFailureComponentRegistrationNotificationEvent");
-
-//        System.out.println("ACTOR ASSET ISSUER - FAILED CONNECTION WITH " + remoteParticipant.getCommunicationCloudClientIdentity() + "\n" +
-//                "--------------------------------------------------------");
-//        assetIssuerActorNetworkServiceAgent.connectionFailure(remoteParticipant.getIdentityPublicKey());
-
-        //I check my time trying to send the message
-//        checkFailedDeliveryTime(remoteParticipant.getIdentityPublicKey());
-
-    }
-
-    @Override
-    public void handleCompleteRequestListComponentRegisteredNotificationEvent(List<PlatformComponentProfile> platformComponentProfileRegisteredList) {
-        System.out.println(" CommunicationNetworkServiceConnectionManager - Starting method handleCompleteComponentConnectionRequestNotificationEvent");
-
-        /*
-         * if have result create a ActorAssetIssuer
-         */
-
-        if (platformComponentProfileRegisteredList != null && !platformComponentProfileRegisteredList.isEmpty()) {
-
-            /*
-             * Get a remote network service registered from the list requested
-             */
-            PlatformComponentProfile remoteNetworkServiceToConnect = platformComponentProfileRegisteredList.get(0);
-
-            /*
-             * tell to the manager to connect to this remote network service
-             */
-            //communicationNetworkServiceConnectionManager.connectTo(remoteNetworkServiceToConnect);
-
-
-            if (remoteNetworkServiceToConnect.getNetworkServiceType() == NetworkServiceType.UNDEFINED && remoteNetworkServiceToConnect.getPlatformComponentType() == PlatformComponentType.ACTOR_ASSET_ISSUER) {
-
-                for (PlatformComponentProfile p : platformComponentProfileRegisteredList) {
-
-                    Location loca = null;
-                    ActorAssetIssuer actorAssetIssuerNew = new AssetIssuerActorRecord(p.getIdentityPublicKey(), p.getName(), convertoByteArrayfromString(p.getExtraData()), loca);
-
-                    actorAssetIssuerRegisteredList.add(actorAssetIssuerNew);
-                }
-
-                FermatEvent event = eventManager.getNewEvent(EventType.COMPLETE_REQUEST_LIST_ASSET_ISSUER_REGISTERED_NOTIFICATION);
-                event.setSource(EventSource.ACTOR_ASSET_ISSUER);
-
-                //((AssetUserActorRequestListRegisteredNetworkServiceNotificationEvent) event).setActorAssetUserList(actorAssetUserRegisteredList);
-                eventManager.raiseEvent(event);
-
-
-            } else if (remoteNetworkServiceToConnect.getNetworkServiceType() == NetworkServiceType.ASSET_ISSUER_ACTOR && remoteNetworkServiceToConnect.getPlatformComponentType() == PlatformComponentType.NETWORK_SERVICE) {
-
-                /*
-                 * save into the cache
-                 */
-                remoteNetworkServicesRegisteredList = platformComponentProfileRegisteredList;
-
+    private void startTimer() {
+        timer.schedule(new TimerTask() {
+            @Override
+            public void run() {
+                // change message state to process retry later
+                reprocessPendingMessage();
             }
-        }
+        }, 0, reprocessTimer);
     }
 
     @Override
-    public void handleCompleteComponentConnectionRequestNotificationEvent(PlatformComponentProfile applicantComponentProfile, PlatformComponentProfile remoteComponentProfile) {
-
-        System.out.println(" TemplateNetworkServiceRoot - Starting method handleCompleteComponentConnectionRequestNotificationEvent");
-
-        /*
-         * Tell the manager to handler the new connection stablished
-         */
-        communicationNetworkServiceConnectionManager.handleEstablishedRequestedNetworkServiceConnection(remoteComponentProfile);
+    public List<DeveloperDatabase> getDatabaseList(DeveloperObjectFactory developerObjectFactory) {
+        return assetIssuerNetworkServiceDeveloperDatabaseFactory.getDatabaseList(developerObjectFactory);
     }
 
-    /**
-     * Handles the events VPNConnectionCloseNotificationEvent
-     *
-     * @param fermatEvent
-     */
     @Override
-    public void handleVpnConnectionCloseNotificationEvent(FermatEvent fermatEvent) {
-
-        if (fermatEvent instanceof VPNConnectionCloseNotificationEvent) {
-
-            VPNConnectionCloseNotificationEvent vpnConnectionCloseNotificationEvent = (VPNConnectionCloseNotificationEvent) fermatEvent;
-
-            if (vpnConnectionCloseNotificationEvent.getNetworkServiceApplicant() == getNetworkServiceType()) {
-                String remotePublicKey = vpnConnectionCloseNotificationEvent.getRemoteParticipant().getIdentityPublicKey();
-
-                if (communicationNetworkServiceConnectionManager != null)
-                    communicationNetworkServiceConnectionManager.closeConnection(remotePublicKey);
-
-                if (assetIssuerActorNetworkServiceAgent != null)
-                    assetIssuerActorNetworkServiceAgent.getPoolConnectionsWaitingForResponse().remove(remotePublicKey);
-
-            }
-        }
+    public List<DeveloperDatabaseTable> getDatabaseTableList(DeveloperObjectFactory developerObjectFactory, DeveloperDatabase developerDatabase) {
+        if(developerDatabase.getName().equals(AssetIssuerNetworkServiceDatabaseConstants.DATA_BASE_NAME))
+            return new AssetIssuerNetworkServiceDeveloperDatabaseFactory(pluginDatabaseSystem, pluginId).getDatabaseTableList(developerObjectFactory);
+        else
+            return new AssetIssuerNetworkServiceDeveloperDatabaseFactory(pluginDatabaseSystem, pluginId).getDatabaseTableListCommunication(developerObjectFactory);
     }
 
-    /**
-     * Handles the events ClientConnectionCloseNotificationEvent
-     *
-     * @param fermatEvent
-     */
     @Override
-    public void handleClientConnectionCloseNotificationEvent(FermatEvent fermatEvent) {
-
-        if (fermatEvent instanceof ClientConnectionCloseNotificationEvent) {
-            this.register = false;
-
-            if (communicationNetworkServiceConnectionManager != null)
-                communicationNetworkServiceConnectionManager.closeAllConnection();
-        }
-
+    public List<DeveloperDatabaseTableRecord> getDatabaseTableContent(DeveloperObjectFactory developerObjectFactory, DeveloperDatabase developerDatabase, DeveloperDatabaseTable developerDatabaseTable) {
+        return assetIssuerNetworkServiceDeveloperDatabaseFactory.getDatabaseTableContent(developerObjectFactory, developerDatabase, developerDatabaseTable);
     }
 
-    /*
-    * Handles the events ClientConnectionLooseNotificationEvent
-    */
     @Override
-    public void handleClientConnectionLooseNotificationEvent(FermatEvent fermatEvent) {
-
-        if (communicationNetworkServiceConnectionManager != null)
-            communicationNetworkServiceConnectionManager.stop();
-
+    public List<String> getClassesFullPath() {
+        return null;
     }
 
-    /*
-     * Handles the events ClientSuccessfullReconnectNotificationEvent
-     */
     @Override
-    public void handleClientSuccessfullReconnectNotificationEvent(FermatEvent fermatEvent) {
-        if (communicationNetworkServiceConnectionManager != null) {
-            communicationNetworkServiceConnectionManager.restart();
-        }
+    public void setLoggingLevelPerClass(Map<String, LogLevel> newLoggingLevel) {
 
-
-        if (communicationRegistrationProcessNetworkServiceAgent != null && !this.register) {
-
-            if (communicationRegistrationProcessNetworkServiceAgent.isAlive()) {
-                communicationRegistrationProcessNetworkServiceAgent.interrupt();
-                communicationRegistrationProcessNetworkServiceAgent = null;
-            }
-
-                /*
-                 * Construct my profile and register me
-                 */
-            PlatformComponentProfile platformComponentProfileToReconnect = wsCommunicationsCloudClientManager.getCommunicationsCloudClientConnection().
-                    constructPlatformComponentProfileFactory(this.getIdentityPublicKey(),
-                            this.getAlias().toLowerCase(),
-                            this.getName(),
-                            this.getNetworkServiceType(),
-                            this.getPlatformComponentType(),
-                            this.getExtraData());
-
-            try {
-                    /*
-                     * Register me
-                     */
-                wsCommunicationsCloudClientManager.getCommunicationsCloudClientConnection().registerComponentForCommunication(
-                        this.getNetworkServiceType(),
-                        platformComponentProfileToReconnect);
-
-            } catch (CantRegisterComponentException e) {
-                e.printStackTrace();
-            }
-                /*
-                 * Configure my new profile
-                 */
-            this.setPlatformComponentProfilePluginRoot(platformComponentProfileToReconnect);
-
-                /*
-                 * Initialize the connection manager
-                 */
-            this.initializeCommunicationNetworkServiceConnectionManager();
-        }
-
-        /*
-         * Mark as register
-         */
-        this.register = Boolean.TRUE;
-
-        if (assetIssuerActorNetworkServiceAgent != null) {
-            try {
-                assetIssuerActorNetworkServiceAgent.start();
-            } catch (CantStartAgentException e) {
-                e.printStackTrace();
-            }
-        } else {
-            initializeActorAssetIssuerAgent();
-        }
     }
 
-    /*
-     * get a string and convert it into array bytes
-     */
-    private static byte[] convertoByteArrayfromString(String arrengebytes) {
+    private void reportUnexpectedError(final Exception e) {
+        errorManager.reportUnexpectedPluginException(Plugins.BITDUBAI_DAP_ASSET_ISSUER_ACTOR_NETWORK_SERVICE, UnexpectedPluginExceptionSeverity.DISABLES_SOME_FUNCTIONALITY_WITHIN_THIS_PLUGIN, e);
+    }
 
-        if (arrengebytes != null) {
-            try {
-                String[] byteValues = arrengebytes.substring(1, arrengebytes.length() - 1).split(",");
-                byte[] bytes = new byte[byteValues.length];
-                if (bytes.length > 0) {
-                    for (int i = 0, len = bytes.length; i < len; i++) {
-                        bytes[i] = Byte.parseByte(byteValues[i].trim());
+    public void buildSendMessage(ActorNotification actorNotification) {
+
+        try {
+            UUID newNotificationID = UUID.randomUUID();
+            AssetNotificationDescriptor assetNotificationDescriptor = AssetNotificationDescriptor.EXTENDED_KEY;
+            long currentTime = System.currentTimeMillis();
+            ActorAssetProtocolState actorAssetProtocolState = ActorAssetProtocolState.PROCESSING_SEND;
+
+            final ActorAssetNetworkServiceRecord assetIssuerNetworkServiceRecord = outgoingNotificationDao.createNotification(
+                    newNotificationID,
+                    actorNotification.getActorDestinationPublicKey(),//redeem
+                    actorNotification.getActorDestinationType(),
+                    actorNotification.getActorSenderPublicKey(),//issuer
+                    actorNotification.getActorSenderAlias(),
+//                    intraUserToAddPhrase,
+                    actorNotification.getActorSenderProfileImage(),//redeem
+                    actorNotification.getActorSenderType(),
+                    assetNotificationDescriptor,
+                    currentTime,
+                    actorAssetProtocolState,
+                    false,
+                    1,
+                    null,
+                    null,
+                    null
+            );
+
+            executorService.submit(new Runnable() {
+                @Override
+                public void run() {
+                    try {
+                        sendNewMessage(
+//                                getProfileSenderToRequestConnection(assetIssuerNetworkServiceRecord.getActorSenderPublicKey()),
+//                                getProfileDestinationToRequestConnection(assetIssuerNetworkServiceRecord.getActorDestinationPublicKey()),
+//                                assetIssuerNetworkServiceRecord.toJson());
+                                getProfileSenderToRequestConnection(
+                                        assetIssuerNetworkServiceRecord.getActorSenderPublicKey(),
+                                        NetworkServiceType.UNDEFINED,
+                                        platformComponentTypeSelectorByActorType(assetIssuerNetworkServiceRecord.getActorSenderType())
+                                ),
+                                getProfileDestinationToRequestConnection(
+                                        assetIssuerNetworkServiceRecord.getActorDestinationPublicKey(),
+                                        NetworkServiceType.UNDEFINED,
+                                        platformComponentTypeSelectorByActorType(assetIssuerNetworkServiceRecord.getActorDestinationType())
+                                ),
+                                assetIssuerNetworkServiceRecord.toJson());
+                    } catch (com.bitdubai.fermat_p2p_api.layer.all_definition.communication.network_services.exceptions.CantSendMessageException | InvalidParameterException e) {
+                        e.printStackTrace();
                     }
-                    return bytes;
-                } else {
-                    return new byte[]{};
                 }
-            } catch (Exception e) {
-                return new byte[]{};
-            }
-        } else {
-            return new byte[]{};
+            });
+            // Sending message to the destination
+        } catch (final CantCreateActorAssetNotificationException e) {
+            reportUnexpectedError(e);
+            e.printStackTrace();
+        }
+    }
+
+    public void responseExtended(ActorNotification actorNotification, ExtendedPublicKey extendedPublicKey) {
+
+        try {
+            UUID newNotificationID = UUID.randomUUID();
+            AssetNotificationDescriptor assetNotificationDescriptor = AssetNotificationDescriptor.EXTENDED_KEY;
+            long currentTime = System.currentTimeMillis();
+            ActorAssetProtocolState actorAssetProtocolState = ActorAssetProtocolState.PROCESSING_SEND;
+
+            String extendedXML = XMLParser.parseObject(extendedPublicKey);
+
+            final ActorAssetNetworkServiceRecord assetIssuerNetworkServiceRecord = outgoingNotificationDao.createNotification(
+                    newNotificationID,
+                    actorNotification.getActorDestinationPublicKey(),//issuer
+                    actorNotification.getActorDestinationType(),
+                    actorNotification.getActorSenderPublicKey(),//redeem
+                    actorNotification.getActorSenderAlias(),
+//                    intraUserToAddPhrase,
+                    actorNotification.getActorSenderProfileImage(),//redeem
+                    actorNotification.getActorSenderType(),
+                    assetNotificationDescriptor,
+                    currentTime,
+                    actorAssetProtocolState,
+                    false,
+                    1,
+                    null,
+                    null,
+                    extendedXML
+            );
+
+            executorService.submit(new Runnable() {
+                @Override
+                public void run() {
+                    try {
+                        sendNewMessage(
+//                                getProfileSenderToRequestConnection(assetIssuerNetworkServiceRecord.getActorSenderPublicKey()),
+//                                getProfileDestinationToRequestConnection(assetIssuerNetworkServiceRecord.getActorDestinationPublicKey()),
+//                                assetIssuerNetworkServiceRecord.toJson());
+                                getProfileSenderToRequestConnection(
+                                        assetIssuerNetworkServiceRecord.getActorSenderPublicKey(),
+                                        NetworkServiceType.UNDEFINED,
+                                        platformComponentTypeSelectorByActorType(assetIssuerNetworkServiceRecord.getActorSenderType())
+                                ),
+                                getProfileDestinationToRequestConnection(
+                                        assetIssuerNetworkServiceRecord.getActorDestinationPublicKey(),
+                                        NetworkServiceType.UNDEFINED,
+                                        platformComponentTypeSelectorByActorType(assetIssuerNetworkServiceRecord.getActorDestinationType())
+                                ),
+                                assetIssuerNetworkServiceRecord.toJson());
+                    } catch (com.bitdubai.fermat_p2p_api.layer.all_definition.communication.network_services.exceptions.CantSendMessageException | InvalidParameterException e) {
+                        e.printStackTrace();
+                    }
+                }
+            });
+            // Sending message to the destination
+        } catch (final CantCreateActorAssetNotificationException e) {
+            reportUnexpectedError(e);
+            e.printStackTrace();
         }
     }
 
@@ -1535,13 +1423,17 @@ public class AssetIssuerActorNetworkServicePluginRoot extends AbstractNetworkSer
 //    public List<FermatMessage> getNewReceivedMessageList() throws CantReadRecordDataBaseException {
 //
 //        Map<String, Object> filters = new HashMap<>();
-//        filters.put(AssetIssuerNetworkServiceDatabaseConstants.INCOMING_MESSAGES_STATUS_COLUMN_NAME, MessagesStatus.NEW_RECEIVED.getCode());
+//        filters.put(CommunicationNetworkServiceDatabaseConstants.INCOMING_MESSAGES_STATUS_COLUMN_NAME, MessagesStatus.NEW_RECEIVED.getCode());
 //
-//        return communicationNetworkServiceConnectionManager.getIncomingMessageDao().findAll(filters);
+//        try {
+//            return getCommunicationNetworkServiceConnectionManager().getIncomingMessageDao().findAll(filters);
+//        } catch (com.bitdubai.fermat_p2p_api.layer.all_definition.communication.network_services.exceptions.CantReadRecordDataBaseException e) {
+//            e.printStackTrace();
+//        }
+//        return null;
 //    }
-
     public List<DAPMessage> getUnreadDAPMessagesByType(DAPMessageType type) throws CantGetDAPMessagesException {
-//        String context = "Message Type: " + type;
+        String context = "Message Type: " + type;
         List<DAPMessage> listToReturn = new ArrayList<>();
 //        try {
 //            for (FermatMessage message : getNewReceivedMessageList()) {
@@ -1556,126 +1448,9 @@ public class AssetIssuerActorNetworkServicePluginRoot extends AbstractNetworkSer
 //                    continue; //This statement is unnecessary but I'll keep it so people can understand better.
 //                }
 //            }
-            return listToReturn;
+        return listToReturn;
 //        } catch (CantReadRecordDataBaseException | CantUpdateRecordDataBaseException e) {
 //            throw new CantGetDAPMessagesException(context, e);
 //        }
-    }
-
-//    @Override
-//    public List<DAPMessage> getUnreadDAPMessageBySubject(DAPMessageSubject subject) throws CantGetDAPMessagesException {
-//        return Collections.EMPTY_LIST;
-//    }
-//
-//    @Override
-//    public void confirmReception(DAPMessage message) throws CantUpdateMessageStatusException {
-//
-//    }
-
-    /**
-     * Mark the message as read
-     *
-     * @param fermatMessage
-     */
-    public void markAsRead(FermatMessage fermatMessage) throws CantUpdateRecordDataBaseException {
-
-        ((FermatMessageCommunication) fermatMessage).setFermatMessagesStatus(FermatMessagesStatus.READ);
-        communicationNetworkServiceConnectionManager.getIncomingMessageDao().update(fermatMessage);
-    }
-
-    @Override
-    public List<DeveloperDatabase> getDatabaseList(DeveloperObjectFactory developerObjectFactory) {
-        return communicationNetworkServiceDeveloperDatabaseFactory.getDatabaseList(developerObjectFactory);
-    }
-
-    @Override
-    public List<DeveloperDatabaseTable> getDatabaseTableList(DeveloperObjectFactory developerObjectFactory, DeveloperDatabase developerDatabase) {
-        return communicationNetworkServiceDeveloperDatabaseFactory.getDatabaseTableList(developerObjectFactory);
-    }
-
-    @Override
-    public List<DeveloperDatabaseTableRecord> getDatabaseTableContent(DeveloperObjectFactory developerObjectFactory, DeveloperDatabase developerDatabase, DeveloperDatabaseTable developerDatabaseTable) {
-        return communicationNetworkServiceDeveloperDatabaseFactory.getDatabaseTableContent(developerObjectFactory, developerDatabase, developerDatabaseTable);
-    }
-
-    /**
-     * Initialize the event listener and configure
-     */
-    private void initializeListener() {
-                 /*
-         * Listen and handle Complete Component Registration Notification Event
-         */
-        FermatEventListener fermatEventListener = eventManager.getNewListener(P2pEventType.COMPLETE_COMPONENT_REGISTRATION_NOTIFICATION);
-        fermatEventListener.setEventHandler(new CompleteComponentRegistrationNotificationEventHandler(this));
-        eventManager.addListener(fermatEventListener);
-        listenersAdded.add(fermatEventListener);
-
-         /*
-         * Listen and handle Complete Request List Component Registered Notification Event
-         */
-        fermatEventListener = eventManager.getNewListener(P2pEventType.COMPLETE_REQUEST_LIST_COMPONENT_REGISTERED_NOTIFICATION);
-        fermatEventListener.setEventHandler(new CompleteRequestListComponentRegisteredNotificationEventHandler(this));
-        eventManager.addListener(fermatEventListener);
-        listenersAdded.add(fermatEventListener);
-
-        /*
-         * Listen and handle Complete Request List Component Registered Notification Event
-         */
-        fermatEventListener = eventManager.getNewListener(P2pEventType.COMPLETE_COMPONENT_CONNECTION_REQUEST_NOTIFICATION);
-        fermatEventListener.setEventHandler(new CompleteComponentConnectionRequestNotificationEventHandler(this));
-        eventManager.addListener(fermatEventListener);
-        listenersAdded.add(fermatEventListener);
-
-        /*
-         * Listen and handle new message sent
-         */
-        fermatEventListener = eventManager.getNewListener(P2pEventType.NEW_NETWORK_SERVICE_MESSAGE_SENT_NOTIFICATION);
-        fermatEventListener.setEventHandler(new NewSentMessagesNotificationEventHandler());
-        eventManager.addListener(fermatEventListener);
-        listenersAdded.add(fermatEventListener);
-
-        /*
-         * Listen and handle New Message Receive Notification Event
-         */
-        fermatEventListener = eventManager.getNewListener(P2pEventType.NEW_NETWORK_SERVICE_MESSAGE_RECEIVE_NOTIFICATION);
-        if (communicationNetworkServiceConnectionManager == null)
-            initializeCommunicationNetworkServiceConnectionManager();
-        fermatEventListener.setEventHandler(new NewReceiveMessagesNotificationEventHandler(communicationNetworkServiceConnectionManager, eventManager));
-        eventManager.addListener(fermatEventListener);
-        listenersAdded.add(fermatEventListener);
-
-        /*
-         * Listen and handle VPN Connection Close Notification Event
-         */
-        fermatEventListener = eventManager.getNewListener(P2pEventType.VPN_CONNECTION_CLOSE);
-        fermatEventListener.setEventHandler(new VPNConnectionCloseNotificationEventHandler(this));
-        eventManager.addListener(fermatEventListener);
-        listenersAdded.add(fermatEventListener);
-
-        /*
-         * Listen and handle Client Connection Close Notification Event
-         */
-        fermatEventListener = eventManager.getNewListener(P2pEventType.CLIENT_CONNECTION_CLOSE);
-        fermatEventListener.setEventHandler(new ClientConnectionCloseNotificationEventHandler(this));
-        eventManager.addListener(fermatEventListener);
-        listenersAdded.add(fermatEventListener);
-
-          /*
-         * Listen and handle Client Connection Loose Notification Event
-         */
-        fermatEventListener = eventManager.getNewListener(P2pEventType.CLIENT_CONNECTION_LOOSE);
-        fermatEventListener.setEventHandler(new ClientConnectionLooseNotificationEventHandler(this));
-        eventManager.addListener(fermatEventListener);
-        listenersAdded.add(fermatEventListener);
-
-
-        /*
-         * Listen and handle Client Connection Success Reconnect Notification Event
-         */
-        fermatEventListener = eventManager.getNewListener(P2pEventType.CLIENT_SUCCESS_RECONNECT);
-        fermatEventListener.setEventHandler(new ClientSuccessfullReconnectNotificationEventHandler(this));
-        eventManager.addListener(fermatEventListener);
-        listenersAdded.add(fermatEventListener);
-
     }
 }
